Split user and role tests
