+++ rest/adminapitest/admin_api_test.go
@@ -9,11 +9,8 @@
 package adminapitest
 
 import (
-	"bufio"
-	"bytes"
 	"encoding/json"
 	"fmt"
-	"io"
 	"log"
 	"net/http"
 	"net/url"
@@ -162,275 +159,6 @@ func TestNoPanicInvalidUpdate(t *testing.T) {
 	revGeneration, _ = db.ParseRevID(revId)
 	assert.Equal(t, 2, revGeneration)
 
-}
-
-func TestUserPasswordValidation(t *testing.T) {
-
-	// PUT a user
-	rt := rest.NewRestTester(t, nil)
-	defer rt.Close()
-
-	response := rt.SendAdminRequest("PUT", "/db/_user/snej", `{"email":"jens@couchbase.com", "password":"letmein", "admin_channels":["foo", "bar"]}`)
-	rest.RequireStatus(t, response, 201)
-
-	// PUT a user without a password, should fail
-	response = rt.SendAdminRequest("PUT", "/db/_user/ajresnopassword", `{"email":"ajres@couchbase.com", "admin_channels":["foo", "bar"]}`)
-	rest.RequireStatus(t, response, 400)
-
-	// POST a user without a password, should fail
-	response = rt.SendAdminRequest("POST", "/db/_user/", `{"name":"ajresnopassword", "email":"ajres@couchbase.com", "admin_channels":["foo", "bar"]}`)
-	rest.RequireStatus(t, response, 400)
-
-	// PUT a user with a two character password, should fail
-	response = rt.SendAdminRequest("PUT", "/db/_user/ajresnopassword", `{"email":"ajres@couchbase.com", "password":"in", "admin_channels":["foo", "bar"]}`)
-	rest.RequireStatus(t, response, 400)
-
-	// POST a user with a two character password, should fail
-	response = rt.SendAdminRequest("POST", "/db/_user/", `{"name":"ajresnopassword", "email":"ajres@couchbase.com", "password":"an", "admin_channels":["foo", "bar"]}`)
-	rest.RequireStatus(t, response, 400)
-
-	// PUT a user with a zero character password, should fail
-	response = rt.SendAdminRequest("PUT", "/db/_user/ajresnopassword", `{"email":"ajres@couchbase.com", "password":"", "admin_channels":["foo", "bar"]}`)
-	rest.RequireStatus(t, response, 400)
-
-	// POST a user with a zero character password, should fail
-	response = rt.SendAdminRequest("POST", "/db/_user/", `{"name":"ajresnopassword", "email":"ajres@couchbase.com", "password":"", "admin_channels":["foo", "bar"]}`)
-	rest.RequireStatus(t, response, 400)
-
-	// PUT update a user with a two character password, should fail
-	response = rt.SendAdminRequest("PUT", "/db/_user/snej", `{"password":"an"}`)
-	rest.RequireStatus(t, response, 400)
-
-	// PUT update a user with a one character password, should fail
-	response = rt.SendAdminRequest("PUT", "/db/_user/snej", `{"password":"a"}`)
-	rest.RequireStatus(t, response, 400)
-
-	// PUT update a user with a zero character password, should fail
-	response = rt.SendAdminRequest("PUT", "/db/_user/snej", `{"password":""}`)
-	rest.RequireStatus(t, response, 400)
-
-	// PUT update a user with a three character password, should succeed
-	response = rt.SendAdminRequest("PUT", "/db/_user/snej", `{"password":"abc"}`)
-	rest.RequireStatus(t, response, 200)
-}
-
-func TestUserAllowEmptyPassword(t *testing.T) {
-
-	// PUT a user
-	rt := rest.NewRestTester(t, &rest.RestTesterConfig{DatabaseConfig: &rest.DatabaseConfig{DbConfig: rest.DbConfig{AllowEmptyPassword: base.BoolPtr(true)}}})
-	defer rt.Close()
-
-	response := rt.SendAdminRequest("PUT", "/db/_user/snej", `{"email":"jens@couchbase.com", "password":"letmein", "admin_channels":["foo", "bar"]}`)
-	rest.RequireStatus(t, response, 201)
-
-	// PUT a user without a password, should succeed
-	response = rt.SendAdminRequest("PUT", "/db/_user/nopassword1", `{"email":"ajres@couchbase.com", "admin_channels":["foo", "bar"]}`)
-	rest.RequireStatus(t, response, 201)
-
-	// POST a user without a password, should succeed
-	response = rt.SendAdminRequest("POST", "/db/_user/", `{"name":"nopassword2", "email":"ajres@couchbase.com", "admin_channels":["foo", "bar"]}`)
-	rest.RequireStatus(t, response, 201)
-
-	// PUT a user with a two character password, should fail
-	response = rt.SendAdminRequest("PUT", "/db/_user/nopassword3", `{"email":"ajres@couchbase.com", "password":"in", "admin_channels":["foo", "bar"]}`)
-	rest.RequireStatus(t, response, 400)
-
-	// POST a user with a two character password, should fail
-	response = rt.SendAdminRequest("POST", "/db/_user/", `{"name":"nopassword4", "email":"ajres@couchbase.com", "password":"an", "admin_channels":["foo", "bar"]}`)
-	rest.RequireStatus(t, response, 400)
-
-	// PUT a user with a zero character password, should succeed
-	response = rt.SendAdminRequest("PUT", "/db/_user/nopassword5", `{"email":"ajres@couchbase.com", "password":"", "admin_channels":["foo", "bar"]}`)
-	rest.RequireStatus(t, response, 201)
-
-	// POST a user with a zero character password, should succeed
-	response = rt.SendAdminRequest("POST", "/db/_user/", `{"name":"nopassword6", "email":"ajres@couchbase.com", "password":"", "admin_channels":["foo", "bar"]}`)
-	rest.RequireStatus(t, response, 201)
-
-	// PUT update a user with a two character password, should fail
-	response = rt.SendAdminRequest("PUT", "/db/_user/snej", `{"password":"an"}`)
-	rest.RequireStatus(t, response, 400)
-
-	// PUT update a user with a one character password, should fail
-	response = rt.SendAdminRequest("PUT", "/db/_user/snej", `{"password":"a"}`)
-	rest.RequireStatus(t, response, 400)
-
-	// PUT update a user with a zero character password, should succeed
-	response = rt.SendAdminRequest("PUT", "/db/_user/snej", `{"password":""}`)
-	rest.RequireStatus(t, response, 200)
-
-	// PUT update a user with a three character password, should succeed
-	response = rt.SendAdminRequest("PUT", "/db/_user/snej", `{"password":"abc"}`)
-	rest.RequireStatus(t, response, 200)
-}
-
-func TestPrincipalForbidUpdatingChannels(t *testing.T) {
-	rt := rest.NewRestTester(t, &rest.RestTesterConfig{})
-	defer rt.Close()
-
-	// Users
-	// PUT admin_channels
-	response := rt.SendAdminRequest("PUT", "/db/_user/snej", `{"email":"jens@couchbase.com", "password":"letmein", "admin_channels":["foo", "bar"]}`)
-	rest.RequireStatus(t, response, 201)
-
-	// PUT all_channels - should fail
-	response = rt.SendAdminRequest("PUT", "/db/_user/snej", `{"email":"jens@couchbase.com", "all_channels":["baz"]}`)
-	rest.RequireStatus(t, response, 400)
-
-	// PUT admin_roles
-	response = rt.SendAdminRequest("PUT", "/db/_user/snej", `{"email":"jens@couchbase.com", "password":"letmein", "admin_roles":["foo", "bar"]}`)
-	rest.RequireStatus(t, response, 200)
-
-	// PUT roles - should fail
-	response = rt.SendAdminRequest("PUT", "/db/_user/snej", `{"email":"jens@couchbase.com", "roles":["baz"]}`)
-	rest.RequireStatus(t, response, 400)
-
-	// Roles
-	// PUT admin_channels
-	response = rt.SendAdminRequest("PUT", "/db/_role/test", `{"admin_channels":["foo", "bar"]}`)
-	rest.RequireStatus(t, response, 201)
-
-	// PUT all_channels - should fail
-	response = rt.SendAdminRequest("PUT", "/db/_role/test", `{"all_channels":["baz"]}`)
-	rest.RequireStatus(t, response, 400)
-}
-
-// Test user access grant while that user has an active changes feed.  (see issue #880)
-func TestUserAccessRace(t *testing.T) {
-
-	base.LongRunningTest(t)
-
-	// This test only runs against Walrus due to known sporadic failures.
-	// See https://github.com/couchbase/sync_gateway/issues/3006
-	if !base.UnitTestUrlIsWalrus() {
-		t.Skip("Skip this test under integration testing")
-	}
-
-	syncFunction := `
-function(doc, oldDoc) {
-  if (doc.type == "list") {
-    channel("list-"+doc._id);
-  } else if (doc.type == "profile") {
-    channel("profiles");
-    var user = doc._id.substring(doc._id.indexOf(":")+1);
-    if (user !== doc.user_id) {
-      throw({forbidden : "profile user_id must match docid"})
-    }
-    requireUser(user);
-    access(user, "profiles");
-    channel('profile-'+user);
-  } else if (doc.type == "Want") {
-    var parts = doc._id.split("-");
-    var user = parts[1];
-    var i = parts[2];
-    requireUser(user);
-    channel('profile-'+user);
-    access(user, 'list-'+i);
-  }
-}
-
-`
-	rtConfig := rest.RestTesterConfig{SyncFn: syncFunction}
-	var rt = rest.NewRestTester(t, &rtConfig)
-	defer rt.Close()
-
-	response := rt.SendAdminRequest("PUT", "/db/_user/bernard", `{"name":"bernard", "password":"letmein", "admin_channels":["profile-bernard"]}`)
-	rest.RequireStatus(t, response, 201)
-
-	// Try to force channel initialisation for user bernard
-	response = rt.SendAdminRequest("GET", "/db/_user/bernard", "")
-	rest.RequireStatus(t, response, 200)
-
-	// Create list docs
-	input := `{"docs": [`
-
-	for i := 1; i <= 100; i++ {
-		if i > 1 {
-			input = input + `,`
-		}
-		docId := fmt.Sprintf("l_%d", i)
-		input = input + fmt.Sprintf(`{"_id":"%s", "type":"list"}`, docId)
-	}
-	input = input + `]}`
-	response = rt.SendAdminRequest("POST", "/db/_bulk_docs", input)
-
-	// Start changes feed
-	var wg sync.WaitGroup
-
-	// Init the public handler, to avoid data race initializing in the two usages below.
-	_ = rt.SendRequest("GET", "/db", "")
-
-	wg.Add(1)
-
-	numExpectedChanges := 201
-
-	go func() {
-		defer wg.Done()
-
-		since := ""
-
-		maxTries := 10
-		numTries := 0
-
-		changesAccumulated := []db.ChangeEntry{}
-
-		for {
-
-			// Timeout allows us to read continuous changes after processing is complete.  Needs to be long enough to
-			// ensure it doesn't terminate before the first change is sent.
-			log.Printf("Invoking _changes?feed=continuous&since=%s&timeout=2000", since)
-			changesResponse := rt.Send(rest.RequestByUser("GET", fmt.Sprintf("/db/_changes?feed=continuous&since=%s&timeout=2000", since), "", "bernard"))
-
-			changes, err := readContinuousChanges(changesResponse)
-			assert.NoError(t, err)
-
-			changesAccumulated = append(changesAccumulated, changes...)
-
-			if len(changesAccumulated) >= numExpectedChanges {
-				log.Printf("Got numExpectedChanges (%d).  Done", numExpectedChanges)
-				break
-			} else {
-				log.Printf("Only received %d out of %d expected changes.  Attempt %d / %d.", len(changesAccumulated), numExpectedChanges, numTries, maxTries)
-			}
-
-			// Advance the since value if we got any changes
-			if len(changes) > 0 {
-				since = changes[len(changes)-1].Seq.String()
-				log.Printf("Setting since value to: %s.", since)
-			}
-
-			numTries++
-			if numTries > maxTries {
-				t.Errorf("Giving up trying to receive %d changes.  Only received %d", numExpectedChanges, len(changesAccumulated))
-				return
-			}
-
-		}
-
-	}()
-
-	// Make bulk docs calls, 100 docs each, all triggering access grants to the list docs.
-	for j := 0; j < 1; j++ {
-
-		input := `{"docs": [`
-		for i := 1; i <= 100; i++ {
-			if i > 1 {
-				input = input + `,`
-			}
-			k := j*100 + i
-			docId := fmt.Sprintf("Want-bernard-l_%d", k)
-			input = input + fmt.Sprintf(`{"_id":"%s", "type":"Want", "owner":"bernard"}`, docId)
-		}
-		input = input + `]}`
-
-		log.Printf("Sending 2nd round of _bulk_docs")
-		response = rt.Send(rest.RequestByUser("POST", "/db/_bulk_docs", input, "bernard"))
-		log.Printf("Sent 2nd round of _bulk_docs")
-
-	}
-
-	// wait for changes feed to complete (time out)
-	wg.Wait()
 }
 
 func TestLoggingKeys(t *testing.T) {
@@ -594,173 +322,6 @@ func TestGetStatus(t *testing.T) {
 	response = rt.SendAdminRequest("OPTIONS", "/_status", "")
 	rest.RequireStatus(t, response, 204)
 	assert.Equal(t, "GET", response.Header().Get("Allow"))
-}
-
-// Test user delete while that user has an active changes feed (see issue 809)
-func TestUserDeleteDuringChangesWithAccess(t *testing.T) {
-
-	base.SetUpTestLogging(t, base.LevelInfo, base.KeyChanges, base.KeyCache, base.KeyHTTP)
-
-	rtConfig := rest.RestTesterConfig{SyncFn: `function(doc) {channel(doc.channel); if(doc.type == "setaccess") { access(doc.owner, doc.channel);}}`}
-	rt := rest.NewRestTester(t, &rtConfig)
-	defer rt.Close()
-
-	response := rt.SendAdminRequest("PUT", "/db/_user/bernard", `{"name":"bernard", "password":"letmein", "admin_channels":["foo"]}`)
-	rest.RequireStatus(t, response, 201)
-
-	var wg sync.WaitGroup
-	wg.Add(1)
-	go func() {
-		defer wg.Done()
-		changesResponse := rt.Send(rest.RequestByUser("GET", "/db/_changes?feed=continuous&since=0&timeout=3000", "", "bernard"))
-		// When testing single threaded, this reproduces the issue described in #809.
-		// When testing multithreaded (-cpu 4 -race), there are three (valid) possibilities"
-		// 1. The DELETE gets processed before the _changes auth completes: this will return 401
-		// 2. The _changes request gets processed before the DELETE: the changes response will be closed when the user is deleted
-		// 3. The DELETE is processed after the _changes auth completes, but before the MultiChangesFeed is instantiated.  The
-		//  changes feed doesn't have a trigger to attempt a reload of the user in this scenario, so will continue until disconnected
-		//  by the client.  This should be fixed more generally (to terminate all active user sessions when the user is deleted, not just
-		//  changes feeds) but that enhancement is too high risk to introduce at this time.  The timeout on changes will terminate the unit
-		//  test.
-		if changesResponse.Code == 401 {
-			// case 1 - ok
-		} else {
-			// case 2 - ensure no error processing the changes response.  The number of entries may vary, depending
-			// on whether the changes loop performed an additional iteration before catching the deleted user.
-			_, err := readContinuousChanges(changesResponse)
-			assert.NoError(t, err)
-		}
-	}()
-
-	// TODO: sleep required to ensure the changes feed iteration starts before the delete gets processed.
-	time.Sleep(500 * time.Millisecond)
-	rt.SendAdminRequest("PUT", "/db/bernard_doc1", `{"type":"setaccess", "owner":"bernard","channel":"foo"}`)
-	rt.SendAdminRequest("DELETE", "/db/_user/bernard", "")
-	rt.SendAdminRequest("PUT", "/db/manny_doc1", `{"type":"setaccess", "owner":"manny","channel":"bar"}`)
-	rt.SendAdminRequest("PUT", "/db/bernard_doc2", `{"type":"general", "channel":"foo"}`)
-
-	// case 3
-	for i := 0; i <= 5; i++ {
-		docId := fmt.Sprintf("/db/bernard_doc%d", i+3)
-		response = rt.SendAdminRequest("PUT", docId, `{"type":"setaccess", "owner":"bernard", "channel":"foo"}`)
-	}
-
-	wg.Wait()
-}
-
-// Reads continuous changes feed response into slice of ChangeEntry
-func readContinuousChanges(response *rest.TestResponse) ([]db.ChangeEntry, error) {
-	var change db.ChangeEntry
-	changes := make([]db.ChangeEntry, 0)
-	reader := bufio.NewReader(response.Body)
-	for {
-		entry, readError := reader.ReadBytes('\n')
-		if readError == io.EOF {
-			// done
-			break
-		}
-		if readError != nil {
-			// unexpected read error
-			return changes, readError
-		}
-		entry = bytes.TrimSpace(entry)
-		if len(entry) > 0 {
-			err := base.JSONUnmarshal(entry, &change)
-			if err != nil {
-				return changes, err
-			}
-			changes = append(changes, change)
-			log.Printf("Got change ==> %v", change)
-		}
-
-	}
-	return changes, nil
-}
-
-func TestRoleAPI(t *testing.T) {
-
-	rt := rest.NewRestTester(t, nil)
-	defer rt.Close()
-
-	// PUT a role
-	rest.RequireStatus(t, rt.SendAdminRequest("GET", "/db/_role/hipster", ""), 404)
-	response := rt.SendAdminRequest("PUT", "/db/_role/hipster", `{"admin_channels":["fedoras", "fixies"]}`)
-	rest.RequireStatus(t, response, 201)
-	response = rt.SendAdminRequest("PUT", "/db/_role/testdeleted", `{"admin_channels":["fedoras", "fixies"]}`)
-	rest.RequireStatus(t, response, 201)
-	rest.RequireStatus(t, rt.SendAdminRequest("DELETE", "/db/_role/testdeleted", ""), 200)
-
-	// GET the role and make sure the result is OK
-	response = rt.SendAdminRequest("GET", "/db/_role/hipster", "")
-	rest.RequireStatus(t, response, 200)
-	var body db.Body
-	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &body))
-	assert.Equal(t, "hipster", body["name"])
-	assert.Equal(t, []interface{}{"fedoras", "fixies"}, body["admin_channels"])
-	assert.Equal(t, nil, body["password"])
-
-	response = rt.SendAdminRequest("GET", "/db/_role/", "")
-	rest.RequireStatus(t, response, 200)
-	assert.Equal(t, `["hipster"]`, response.Body.String())
-
-	// DELETE the role
-	rest.RequireStatus(t, rt.SendAdminRequest("DELETE", "/db/_role/hipster", ""), 200)
-	rest.RequireStatus(t, rt.SendAdminRequest("GET", "/db/_role/hipster", ""), 404)
-
-	// POST a role
-	response = rt.SendAdminRequest("POST", "/db/_role", `{"name":"hipster", "admin_channels":["fedoras", "fixies"]}`)
-	rest.RequireStatus(t, response, 301)
-	response = rt.SendAdminRequest("POST", "/db/_role/", `{"name":"hipster", "admin_channels":["fedoras", "fixies"]}`)
-	rest.RequireStatus(t, response, 201)
-	response = rt.SendAdminRequest("GET", "/db/_role/hipster", "")
-	rest.RequireStatus(t, response, 200)
-	body = nil
-	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &body))
-	assert.Equal(t, "hipster", body["name"])
-	rest.RequireStatus(t, rt.SendAdminRequest("DELETE", "/db/_role/hipster", ""), 200)
-
-	// GET including deleted
-	response = rt.SendAdminRequest("GET", "/db/_role/?deleted=true", "")
-	rest.RequireStatus(t, response, 200)
-	assert.Equal(t, `["hipster","testdeleted"]`, response.Body.String())
-}
-
-func TestGuestUser(t *testing.T) {
-
-	guestUserEndpoint := fmt.Sprintf("/db/_user/%s", base.GuestUsername)
-
-	rt := rest.NewRestTester(t, nil)
-	defer rt.Close()
-
-	response := rt.SendAdminRequest(http.MethodGet, guestUserEndpoint, "")
-	rest.RequireStatus(t, response, http.StatusOK)
-	var body db.Body
-	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &body))
-	assert.Equal(t, base.GuestUsername, body["name"])
-	// This ain't no admin-party, this ain't no nightclub, this ain't no fooling around:
-	assert.Nil(t, body["admin_channels"])
-
-	// Disable the guest user:
-	response = rt.SendAdminRequest(http.MethodPut, guestUserEndpoint, `{"disabled":true}`)
-	rest.RequireStatus(t, response, http.StatusOK)
-
-	// Get guest user and verify it is now disabled:
-	response = rt.SendAdminRequest(http.MethodGet, guestUserEndpoint, "")
-	rest.RequireStatus(t, response, http.StatusOK)
-	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &body))
-	assert.Equal(t, base.GuestUsername, body["name"])
-	assert.True(t, body["disabled"].(bool))
-
-	// Check that the actual User object is correct:
-	ctx := rt.Context()
-	user, _ := rt.ServerContext().Database(ctx, "db").Authenticator(ctx).GetUser("")
-	assert.Empty(t, user.Name())
-	assert.Nil(t, user.ExplicitChannels())
-	assert.True(t, user.Disabled())
-
-	// We can't delete the guest user, but we should get a reasonable error back.
-	response = rt.SendAdminRequest(http.MethodDelete, guestUserEndpoint, "")
-	rest.RequireStatus(t, response, http.StatusMethodNotAllowed)
 }
 
 func TestFlush(t *testing.T) {
@@ -2205,142 +1766,6 @@ func TestHandleSGCollect(t *testing.T) {
 	rest.RequireStatus(t, resp, http.StatusBadRequest)
 }
 
-func TestUserAndRoleResponseContentType(t *testing.T) {
-	rt := rest.NewRestTester(t, nil)
-	defer rt.Close()
-
-	// Create a user 'christopher' through PUT request with empty request body.
-	var responseBody db.Body
-	body := `{"email":"christopher@couchbase.com","password":"cGFzc3dvcmQ=","admin_channels":["foo", "bar"]}`
-	response := rt.SendAdminRequest(http.MethodPut, "/db/_user/christopher", "")
-	assert.Equal(t, http.StatusBadRequest, response.Code)
-	assert.Equal(t, "application/json", response.Header().Get("Content-Type"))
-	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &responseBody))
-
-	// Create a user 'charles' through POST request with empty request body.
-	body = `{"email":"charles@couchbase.com","password":"cGFzc3dvcmQ=","admin_channels":["foo", "bar"]}`
-	response = rt.SendAdminRequest(http.MethodPost, "/db/_user/charles", "")
-	assert.Equal(t, http.StatusMethodNotAllowed, response.Code)
-	assert.Equal(t, "application/json", response.Header().Get("Content-Type"))
-	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &responseBody))
-
-	// Create a user 'alice' through PUT request.
-	body = `{"email":"alice@couchbase.com","password":"cGFzc3dvcmQ=","admin_channels":["foo", "bar"]}`
-	response = rt.SendAdminRequest(http.MethodPut, "/db/_user/alice", body)
-	assert.Equal(t, http.StatusCreated, response.Code)
-	assert.Empty(t, response.Header().Get("Content-Type"))
-
-	// Create another user 'bob' through POST request.
-	body = `{"name":"bob","email":"bob@couchbase.com","password":"cGFzc3dvcmQ=","admin_channels":["foo", "bar"]}`
-	response = rt.SendAdminRequest(http.MethodPost, "/db/_user/", body)
-	assert.Equal(t, http.StatusCreated, response.Code)
-	assert.Empty(t, response.Header().Get("Content-Type"))
-
-	// Get the user details of user 'alice' through GET request.
-	response = rt.SendAdminRequest(http.MethodGet, "/db/_user/alice", "")
-	assert.Equal(t, http.StatusOK, response.Code)
-	assert.Equal(t, "application/json", response.Header().Get("Content-Type"))
-	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &responseBody))
-
-	// Get the list of users through GET request.
-	var users []string
-	response = rt.SendAdminRequest(http.MethodGet, "/db/_user/", "")
-	assert.Equal(t, http.StatusOK, response.Code)
-	assert.Equal(t, "application/json", response.Header().Get("Content-Type"))
-	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &users))
-	assert.Subset(t, []string{"alice", "bob"}, users)
-
-	// Check whether the /db/_user/bob resource exist on the server.
-	response = rt.SendAdminRequest(http.MethodHead, "/db/_user/bob", "")
-	assert.Equal(t, http.StatusOK, response.Code)
-	assert.Empty(t, response.Header().Get("Content-Type"))
-
-	// Get the list of users through HEAD request.
-	response = rt.SendAdminRequest(http.MethodHead, "/db/_user/", "")
-	assert.Equal(t, http.StatusOK, response.Code)
-	assert.Empty(t, response.Header().Get("Content-Type"))
-
-	// Delete user 'alice'
-	response = rt.SendAdminRequest(http.MethodDelete, "/db/_user/alice", "")
-	assert.Equal(t, http.StatusOK, response.Code)
-	assert.Empty(t, response.Header().Get("Content-Type"))
-
-	// Delete GUEST user instead of disabling.
-	response = rt.SendAdminRequest(http.MethodDelete, "/db/_user/GUEST", "")
-	assert.Equal(t, http.StatusMethodNotAllowed, response.Code)
-	assert.Equal(t, "application/json", response.Header().Get("Content-Type"))
-
-	// Delete user 'eve' who doesn't exists at this point of time.
-	response = rt.SendAdminRequest(http.MethodDelete, "/db/_user/eve", "")
-	assert.Equal(t, http.StatusNotFound, response.Code)
-	assert.Equal(t, "application/json", response.Header().Get("Content-Type"))
-
-	// Create a new user and save to database to create user session.
-	authenticator := auth.NewAuthenticator(rt.Bucket(), nil, auth.DefaultAuthenticatorOptions())
-	user, err := authenticator.NewUser("eve", "cGFzc3dvcmQ=", channels.BaseSetOf(t, "*"))
-	assert.NoError(t, err, "Couldn't create new user")
-	assert.NoError(t, authenticator.Save(user), "Couldn't save new user")
-
-	// Create user session to check delete session request.
-	response = rt.SendAdminRequest(http.MethodPost, "/db/_session", `{"name":"eve"}`)
-	assert.Equal(t, http.StatusOK, response.Code)
-	assert.Equal(t, "application/json", response.Header().Get("Content-Type"))
-	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &responseBody))
-	sessionId, _ := responseBody["session_id"].(string)
-	require.NotEmpty(t, sessionId, "Couldn't parse sessionID from response body")
-
-	// Delete user session using /db/_user/eve/_session/{sessionId}.
-	response = rt.SendAdminRequest(http.MethodDelete, "/db/_user/eve/_session/"+sessionId, "")
-	assert.Equal(t, http.StatusOK, response.Code)
-	assert.Empty(t, response.Header().Get("Content-Type"))
-
-	// Create user session to check delete session request.
-	response = rt.SendAdminRequest(http.MethodPost, "/db/_session", `{"name":"eve"}`)
-	assert.Equal(t, http.StatusOK, response.Code)
-	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &responseBody))
-
-	// Delete user session using /db/_user/eve/_session request.
-	response = rt.SendAdminRequest(http.MethodDelete, "/db/_user/eve/_session", "")
-	assert.Equal(t, http.StatusOK, response.Code)
-	assert.Empty(t, response.Header().Get("Content-Type"))
-
-	// Create a role 'developer' through POST request
-	body = `{"name":"developer","admin_channels":["channel1", "channel2"]}`
-	response = rt.SendAdminRequest(http.MethodPost, "/db/_role/", body)
-	assert.Equal(t, http.StatusCreated, response.Code)
-	assert.Empty(t, response.Header().Get("Content-Type"))
-
-	// Create another role 'coder' through PUT request.
-	body = `{"admin_channels":["channel3", "channel4"]}`
-	response = rt.SendAdminRequest(http.MethodPut, "/db/_role/coder", body)
-	assert.Equal(t, http.StatusCreated, response.Code)
-	assert.Empty(t, response.Header().Get("Content-Type"))
-
-	// Check whether the /db/_role/ resource exist on the server.
-	response = rt.SendAdminRequest(http.MethodHead, "/db/_role/", "")
-	assert.Equal(t, http.StatusOK, response.Code)
-	assert.Empty(t, response.Header().Get("Content-Type"))
-
-	// Get the created roles through GET request.
-	var roles []string
-	response = rt.SendAdminRequest(http.MethodGet, "/db/_role/", "")
-	assert.Equal(t, http.StatusOK, response.Code)
-	assert.Equal(t, "application/json", response.Header().Get("Content-Type"))
-	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &roles))
-	assert.Subset(t, []string{"coder", "developer"}, roles)
-
-	// Delete role 'coder' from database.
-	response = rt.SendAdminRequest(http.MethodDelete, "/db/_role/coder", "")
-	assert.Equal(t, http.StatusOK, response.Code)
-	assert.Empty(t, response.Header().Get("Content-Type"))
-
-	// Delete role who doesn't exist.
-	response = rt.SendAdminRequest(http.MethodDelete, "/db/_role/programmer", "")
-	assert.Equal(t, http.StatusNotFound, response.Code)
-	assert.Equal(t, "application/json", response.Header().Get("Content-Type"))
-	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &responseBody))
-}
-
 func TestConfigRedaction(t *testing.T) {
 	rt := rest.NewRestTester(t, &rest.RestTesterConfig{DatabaseConfig: &rest.DatabaseConfig{DbConfig: rest.DbConfig{Users: map[string]*auth.PrincipalConfig{"alice": {Password: base.StringPtr("password")}}}}})
 	defer rt.Close()
@@ -2361,97 +1786,6 @@ func TestConfigRedaction(t *testing.T) {
 	require.NoError(t, err)
 
 	assert.Equal(t, base.RedactedStr, unmarshaledServerConfig.Bootstrap.Password)
-}
-
-func TestUserXattrsRawGet(t *testing.T) {
-	if !base.TestUseXattrs() {
-		t.Skip("Test requires xattrs to be enabled")
-	}
-
-	if !base.IsEnterpriseEdition() {
-		t.Skipf("test is EE only - user xattrs")
-	}
-
-	docKey := t.Name()
-	xattrKey := "xattrKey"
-
-	rt := rest.NewRestTester(t, &rest.RestTesterConfig{
-		DatabaseConfig: &rest.DatabaseConfig{
-			DbConfig: rest.DbConfig{
-				AutoImport:   true,
-				UserXattrKey: xattrKey,
-			},
-		},
-	})
-	defer rt.Close()
-
-	userXattrStore, ok := base.AsUserXattrStore(rt.Bucket())
-	if !ok {
-		t.Skip("Test requires Couchbase Bucket")
-	}
-
-	resp := rt.SendAdminRequest("PUT", "/db/"+docKey, "{}")
-	rest.RequireStatus(t, resp, http.StatusCreated)
-	require.NoError(t, rt.WaitForPendingChanges())
-
-	_, err := userXattrStore.WriteUserXattr(docKey, xattrKey, "val")
-	assert.NoError(t, err)
-
-	err = rt.WaitForCondition(func() bool {
-		return rt.GetDatabase().DbStats.SharedBucketImportStats.ImportCount.Value() == 1
-	})
-
-	resp = rt.SendAdminRequest("GET", "/db/_raw/"+docKey, "")
-	rest.RequireStatus(t, resp, http.StatusOK)
-
-	var RawReturn struct {
-		Meta struct {
-			Xattrs map[string]interface{} `json:"xattrs"`
-		} `json:"_meta"`
-	}
-
-	err = json.Unmarshal(resp.BodyBytes(), &RawReturn)
-
-	assert.Equal(t, "val", RawReturn.Meta.Xattrs[xattrKey])
-}
-
-func TestRolePurge(t *testing.T) {
-	rt := rest.NewRestTester(t, nil)
-	defer rt.Close()
-
-	// Create role
-	resp := rt.SendAdminRequest("PUT", "/db/_role/role", `{"admin_channels":["channel"]}`)
-	rest.RequireStatus(t, resp, http.StatusCreated)
-
-	// Delete role
-	resp = rt.SendAdminRequest("DELETE", "/db/_role/role", ``)
-	rest.RequireStatus(t, resp, http.StatusOK)
-
-	// Ensure role is gone
-	resp = rt.SendAdminRequest("GET", "/db/_role/role", ``)
-	rest.RequireStatus(t, resp, http.StatusNotFound)
-
-	// Ensure role is 'soft-deleted' and we can still get the doc
-	role, err := rt.GetDatabase().Authenticator(base.TestCtx(t)).GetRoleIncDeleted("role")
-	assert.NoError(t, err)
-	assert.NotNil(t, role)
-
-	// Re-create role
-	resp = rt.SendAdminRequest("PUT", "/db/_role/role", `{"admin_channels":["channel"]}`)
-	rest.RequireStatus(t, resp, http.StatusCreated)
-
-	// Delete role again but with purge flag
-	resp = rt.SendAdminRequest("DELETE", "/db/_role/role?purge=true", ``)
-	rest.RequireStatus(t, resp, http.StatusOK)
-
-	// Ensure role is purged, can't access at all
-	role, err = rt.GetDatabase().Authenticator(base.TestCtx(t)).GetRoleIncDeleted("role")
-	assert.Nil(t, err)
-	assert.Nil(t, role)
-
-	// Ensure role returns 404 via REST call
-	resp = rt.SendAdminRequest("GET", "/db/_role/role", ``)
-	rest.RequireStatus(t, resp, http.StatusNotFound)
 }
 
 func TestSoftDeleteCasMismatch(t *testing.T) {
@@ -2482,90 +1816,6 @@ func TestSoftDeleteCasMismatch(t *testing.T) {
 
 	resp = rt.SendAdminRequest("PUT", "/db/_role/role", `{"admin_channels":["chan"]}`)
 	rest.RequireStatus(t, resp, http.StatusCreated)
-}
-
-func TestObtainUserChannelsForDeletedRoleCasFail(t *testing.T) {
-	if !base.UnitTestUrlIsWalrus() {
-		t.Skip("Skip LeakyBucket test when running in integration")
-	}
-
-	testCases := []struct {
-		Name      string
-		RunBefore bool
-	}{
-		{
-			"Delete On GetUser",
-			true,
-		},
-		{
-			"Delete On InheritedChannels",
-			false,
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.Name, func(t *testing.T) {
-			rt := rest.NewRestTester(t, &rest.RestTesterConfig{
-				SyncFn: `
-			function(doc, oldDoc){
-				if (doc._id === 'roleChannels'){
-					access('role:role', doc.channels)
-				}
-				if (doc._id === 'userRoles'){
-					role('user', doc.roles)
-				}
-			}
-		`,
-			})
-			defer rt.Close()
-
-			// Create role
-			resp := rt.SendAdminRequest("PUT", "/db/_role/role", `{"admin_channels":["channel"]}`)
-			rest.RequireStatus(t, resp, http.StatusCreated)
-
-			// Create user
-			resp = rt.SendAdminRequest("PUT", "/db/_user/user", `{"password": "pass"}`)
-			rest.RequireStatus(t, resp, http.StatusCreated)
-
-			// Add channel to role
-			resp = rt.SendAdminRequest("PUT", "/db/roleChannels", `{"channels": "inherit"}`)
-			rest.RequireStatus(t, resp, http.StatusCreated)
-
-			// Add role to user
-			resp = rt.SendAdminRequest("PUT", "/db/userRoles", `{"roles": "role:role"}`)
-			rest.RequireStatus(t, resp, http.StatusCreated)
-
-			leakyBucket, ok := base.AsLeakyBucket(rt.TestBucket)
-			require.True(t, ok)
-
-			triggerCallback := false
-			leakyBucket.SetUpdateCallback(func(key string) {
-				if triggerCallback {
-					triggerCallback = false
-					resp = rt.SendAdminRequest("DELETE", "/db/_role/role", ``)
-					rest.RequireStatus(t, resp, http.StatusOK)
-				}
-			})
-
-			if testCase.RunBefore {
-				triggerCallback = true
-			}
-
-			authenticator := rt.GetDatabase().Authenticator(base.TestCtx(t))
-			user, err := authenticator.GetUser("user")
-			assert.NoError(t, err)
-
-			if !testCase.RunBefore {
-				triggerCallback = true
-			}
-
-			assert.Equal(t, []string{"!"}, user.InheritedChannels().AllKeys())
-
-			// Ensure callback ran
-			assert.False(t, triggerCallback)
-		})
-
-	}
 }
 
 // Test warnings being issued when a new channel is created with over 250 characters - CBG-1475
