+++ rest/api_test.go
@@ -21,7 +21,6 @@ import (
 	"mime/multipart"
 	"net/http"
 	"net/http/httptest"
-	"net/url"
 	"os"
 	"path/filepath"
 	"sort"
@@ -207,118 +206,6 @@ func TestFunkyDocIDs(t *testing.T) {
 	RequireStatus(t, response, 201)
 	response = rt.SendAdminRequest("GET", "/db/foo+bar%2Bmoo+car3", "")
 	RequireStatus(t, response, 200)
-}
-
-func TestFunkyUsernames(t *testing.T) {
-	cases := []struct {
-		Name     string
-		UserName string
-	}{
-		{
-			Name:     "hashes",
-			UserName: "foo#bar",
-		},
-		{
-			Name:     "question mark",
-			UserName: "foo?bar",
-		},
-		{
-			Name:     "dollars",
-			UserName: "$foo$bar",
-		},
-		{
-			Name:     "underscore prefix",
-			UserName: "_sync-foobar",
-		},
-	}
-	for _, tc := range cases {
-		t.Run(tc.Name, func(t *testing.T) {
-			require.Truef(t, auth.IsValidPrincipalName(tc.UserName), "expected '%s' to be accepted", tc.UserName)
-			rt := NewRestTester(t, nil)
-			defer rt.Close()
-
-			ctx := rt.Context()
-			a := rt.ServerContext().Database(ctx, "db").Authenticator(ctx)
-
-			// Create a test user
-			user, err := a.NewUser(tc.UserName, "letmein", channels.BaseSetOf(t, "foo"))
-			require.NoError(t, err)
-			require.NoError(t, a.Save(user))
-
-			response := rt.Send(RequestByUser("PUT", "/db/AC+DC", `{"foo":"bar", "channels": ["foo"]}`, tc.UserName))
-			RequireStatus(t, response, 201)
-
-			response = rt.Send(RequestByUser("GET", "/db/_all_docs", "", tc.UserName))
-			RequireStatus(t, response, 200)
-
-			response = rt.Send(RequestByUser("GET", "/db/AC+DC", "", tc.UserName))
-			RequireStatus(t, response, 200)
-			var overlay struct {
-				Rev string `json:"_rev"`
-			}
-			require.NoError(t, json.Unmarshal(response.Body.Bytes(), &overlay))
-
-			response = rt.Send(RequestByUser("DELETE", "/db/AC+DC?rev="+overlay.Rev, "", tc.UserName))
-			RequireStatus(t, response, 200)
-		})
-	}
-}
-
-func TestFunkyRoleNames(t *testing.T) {
-	cases := []struct {
-		Name     string
-		RoleName string
-	}{
-		{
-			Name:     "hashes",
-			RoleName: "foo#bar",
-		},
-		{
-			Name:     "question mark",
-			RoleName: "foo?bar",
-		},
-		{
-			Name:     "dollars",
-			RoleName: "$foo$bar",
-		},
-		{
-			Name:     "underscore prefix",
-			RoleName: "_sync-foobar",
-		},
-	}
-	for _, tc := range cases {
-		t.Run(tc.Name, func(t *testing.T) {
-			require.Truef(t, auth.IsValidPrincipalName(tc.RoleName), "expected '%s' to be accepted", tc.RoleName)
-			roleNameJSON, err := json.Marshal("role:" + tc.RoleName)
-			require.NoError(t, err)
-			const username = "user1"
-			syncFn := fmt.Sprintf(`function(doc) {channel(doc.channels); role("%s", %s);}`, username, string(roleNameJSON))
-			rt := NewRestTester(t, &RestTesterConfig{
-				SyncFn: syncFn,
-			})
-			defer rt.Close()
-
-			ctx := rt.Context()
-			a := rt.ServerContext().Database(ctx, "db").Authenticator(ctx)
-
-			// Create a test user
-			user, err := a.NewUser(username, "letmein", channels.BaseSetOf(t))
-			require.NoError(t, err)
-			require.NoError(t, a.Save(user))
-
-			// Create role
-			response := rt.SendAdminRequest("PUT", fmt.Sprintf("/db/_role/%s", url.PathEscape(tc.RoleName)), `{"admin_channels": ["testchannel"]}`)
-			RequireStatus(t, response, 201)
-
-			// Create test document
-			response = rt.SendAdminRequest("PUT", "/db/testdoc", `{"channels":["testchannel"]}`)
-			RequireStatus(t, response, 201)
-
-			// Assert user can access it
-			response = rt.Send(RequestByUser("GET", "/db/testdoc", "", username))
-			RequireStatus(t, response, 200)
-		})
-	}
 }
 
 func TestCORSOrigin(t *testing.T) {
@@ -901,60 +788,6 @@ func TestBulkDocsChangeToAccess(t *testing.T) {
 	assert.Equal(t, map[string]interface{}{"rev": "1-afbcffa8a4641a0f4dd94d3fc9593e74", "id": "bulk1"}, docs[0])
 
 	assert.Equal(t, map[string]interface{}{"rev": "1-4d79588b9fe9c38faae61f0c1b9471c0", "id": "bulk2"}, docs[1])
-
-}
-
-func TestBulkDocsChangeToRoleAccess(t *testing.T) {
-
-	base.SetUpTestLogging(t, base.LevelDebug, base.KeyAccess)
-
-	rtConfig := RestTesterConfig{SyncFn: `
-		function(doc) {
-			if(doc.type == "roleaccess") {
-				channel(doc.channel);
-				access(doc.grantTo, doc.grantChannel);
-			} else {
-				requireAccess(doc.mustHaveAccess)
-			}
-		}`}
-	rt := NewRestTester(t, &rtConfig)
-	defer rt.Close()
-	ctx := rt.Context()
-
-	// Create a role with no channels assigned to it
-	authenticator := rt.ServerContext().Database(ctx, "db").Authenticator(ctx)
-	role, err := authenticator.NewRole("role1", nil)
-	assert.NoError(t, authenticator.Save(role))
-
-	// Create a user with an explicit role grant for role1
-	user, err := authenticator.NewUser("user1", "letmein", nil)
-	user.SetExplicitRoles(channels.TimedSet{"role1": channels.NewVbSimpleSequence(1)}, 1)
-	err = authenticator.Save(user)
-	assert.NoError(t, err)
-
-	// Bulk docs with 2 docs.  First doc grants role1 access to chan1.  Second requires chan1 for write.
-	input := `{"docs": [
-				{
-					"_id": "bulk1",
-				 	"type" : "roleaccess",
-				 	"grantTo":"role:role1",
-				 	"grantChannel":"chan1"
-				 },
-				{
-					"_id": "bulk2",
-					"mustHaveAccess":"chan1"
-				}
-				]}`
-
-	response := rt.Send(RequestByUser("POST", "/db/_bulk_docs", input, "user1"))
-	RequireStatus(t, response, 201)
-
-	var docs []interface{}
-	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &docs))
-	assert.Equal(t, 2, len(docs))
-	assert.Equal(t, map[string]interface{}{"rev": "1-17424d2a21bf113768dfdbcd344741ac", "id": "bulk1"}, docs[0])
-
-	assert.Equal(t, map[string]interface{}{"rev": "1-f120ccb33c0a6ef43ef202ade28f98ef", "id": "bulk2"}, docs[1])
 
 }
 
@@ -2060,205 +1893,6 @@ func TestSyncFnDocBodyPropertiesSwitchActiveTombstone(t *testing.T) {
 	assert.NoError(t, err)
 
 	assert.Equal(t, 1, numErrorsAfter-numErrorsBefore, "expecting to see only only 1 error logged")
-}
-
-func TestRoleAssignmentBeforeUserExists(t *testing.T) {
-
-	base.SetUpTestLogging(t, base.LevelDebug, base.KeyAccess, base.KeyCRUD, base.KeyChanges)
-
-	rtConfig := RestTesterConfig{SyncFn: `function(doc) {role(doc.user, doc.role);channel(doc.channel)}`}
-	rt := NewRestTester(t, &rtConfig)
-	defer rt.Close()
-
-	ctx := rt.Context()
-	a := rt.ServerContext().Database(ctx, "db").Authenticator(ctx)
-	guest, err := a.GetUser("")
-	assert.NoError(t, err)
-	guest.SetDisabled(false)
-	err = a.Save(guest)
-	assert.NoError(t, err)
-
-	// POST a role
-	response := rt.SendAdminRequest("POST", "/db/_role/", `{"name":"role1","admin_channels":["chan1"]}`)
-	RequireStatus(t, response, 201)
-	response = rt.SendAdminRequest("GET", "/db/_role/role1", "")
-	RequireStatus(t, response, 200)
-	var body db.Body
-	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &body))
-	assert.Equal(t, "role1", body["name"])
-
-	// Put document to trigger sync function
-	response = rt.Send(Request("PUT", "/db/doc1", `{"user":"user1", "role":"role:role1", "channel":"chan1"}`)) // seq=1
-	RequireStatus(t, response, 201)
-	body = nil
-	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &body))
-	assert.Equal(t, true, body["ok"])
-
-	// POST the new user the GET and verify that it shows the assigned role
-	response = rt.SendAdminRequest("POST", "/db/_user/", `{"name":"user1", "password":"letmein"}`)
-	RequireStatus(t, response, 201)
-	response = rt.SendAdminRequest("GET", "/db/_user/user1", "")
-	RequireStatus(t, response, 200)
-	body = nil
-	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &body))
-	assert.Equal(t, "user1", body["name"])
-	assert.Equal(t, []interface{}{"role1"}, body["roles"])
-	assert.Equal(t, []interface{}{"!", "chan1"}, body["all_channels"])
-
-	// goassert.DeepEquals(t, body["admin_roles"], []interface{}{"hipster"})
-	// goassert.DeepEquals(t, body["all_channels"], []interface{}{"bar", "fedoras", "fixies", "foo"})
-}
-
-func TestRoleAccessChanges(t *testing.T) {
-
-	base.SetUpTestLogging(t, base.LevelDebug, base.KeyAccess, base.KeyCRUD, base.KeyChanges)
-
-	rtConfig := RestTesterConfig{SyncFn: `function(doc) {role(doc.user, doc.role);channel(doc.channel)}`}
-	rt := NewRestTester(t, &rtConfig)
-	defer rt.Close()
-
-	ctx := rt.Context()
-	a := rt.ServerContext().Database(ctx, "db").Authenticator(ctx)
-	guest, err := a.GetUser("")
-	require.NoError(t, err)
-
-	guest.SetDisabled(false)
-	err = a.Save(guest)
-	assert.NoError(t, err)
-
-	// Create users:
-	response := rt.SendAdminRequest("PUT", "/db/_user/alice", `{"password":"letmein", "admin_channels":["alpha"]}`)
-	RequireStatus(t, response, 201)
-
-	response = rt.SendAdminRequest("PUT", "/db/_user/zegpold", `{"password":"letmein", "admin_channels":["beta"]}`)
-	RequireStatus(t, response, 201)
-
-	response = rt.SendAdminRequest("PUT", "/db/_role/hipster", `{"admin_channels":["gamma"]}`)
-	RequireStatus(t, response, 201)
-	/*
-		alice, err := a.NewUser("alice", "letmein", channels.BaseSetOf(t, "alpha"))
-		assert.NoError(t, a.Save(alice))
-		zegpold, err := a.NewUser("zegpold", "letmein", channels.BaseSetOf(t, "beta"))
-		assert.NoError(t, a.Save(zegpold))
-
-		hipster, err := a.NewRole("hipster", channels.BaseSetOf(t, "gamma"))
-		assert.NoError(t, a.Save(hipster))
-	*/
-
-	// Create some docs in the channels:
-	cacheWaiter := rt.ServerContext().Database(ctx, "db").NewDCPCachingCountWaiter(t)
-	cacheWaiter.Add(1)
-	response = rt.Send(Request("PUT", "/db/fashion",
-		`{"user":"alice","role":["role:hipster","role:bogus"]}`))
-	RequireStatus(t, response, 201)
-	var body db.Body
-	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &body))
-	assert.Equal(t, true, body["ok"])
-	fashionRevID := body["rev"].(string)
-
-	roleGrantSequence := rt.GetDocumentSequence("fashion")
-
-	cacheWaiter.Add(4)
-	RequireStatus(t, rt.Send(Request("PUT", "/db/g1", `{"channel":"gamma"}`)), 201)
-	RequireStatus(t, rt.Send(Request("PUT", "/db/a1", `{"channel":"alpha"}`)), 201)
-	RequireStatus(t, rt.Send(Request("PUT", "/db/b1", `{"channel":"beta"}`)), 201)
-	RequireStatus(t, rt.Send(Request("PUT", "/db/d1", `{"channel":"delta"}`)), 201)
-
-	// Check user access:
-	alice, _ := a.GetUser("alice")
-	assert.Equal(t,
-
-		channels.TimedSet{
-			"!":     channels.NewVbSimpleSequence(1),
-			"alpha": channels.NewVbSimpleSequence(alice.Sequence()),
-			"gamma": channels.NewVbSimpleSequence(roleGrantSequence),
-		}, alice.InheritedChannels())
-
-	assert.Equal(t,
-
-		channels.TimedSet{
-			"bogus":   channels.NewVbSimpleSequence(roleGrantSequence),
-			"hipster": channels.NewVbSimpleSequence(roleGrantSequence),
-		}, alice.RoleNames())
-
-	zegpold, _ := a.GetUser("zegpold")
-	assert.Equal(t,
-
-		channels.TimedSet{
-			"!":    channels.NewVbSimpleSequence(1),
-			"beta": channels.NewVbSimpleSequence(zegpold.Sequence()),
-		}, zegpold.InheritedChannels())
-
-	assert.Equal(t, channels.TimedSet{}, zegpold.RoleNames())
-
-	// Check the _changes feed:
-	var changes struct {
-		Results  []db.ChangeEntry
-		Last_Seq interface{}
-	}
-
-	cacheWaiter.Wait()
-	response = rt.Send(RequestByUser("GET", "/db/_changes", "", "alice"))
-	log.Printf("1st _changes looks like: %s", response.Body.Bytes())
-	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &changes))
-	require.Equal(t, 3, len(changes.Results))
-	assert.Equal(t, "_user/alice", changes.Results[0].ID)
-	assert.Equal(t, "g1", changes.Results[1].ID)
-	assert.Equal(t, "a1", changes.Results[2].ID)
-
-	response = rt.Send(RequestByUser("GET", "/db/_changes", "", "zegpold"))
-	log.Printf("2nd _changes looks like: %s", response.Body.Bytes())
-	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &changes))
-	require.Equal(t, 2, len(changes.Results))
-	assert.Equal(t, "_user/zegpold", changes.Results[0].ID)
-	assert.Equal(t, "b1", changes.Results[1].ID)
-	lastSeqPreGrant := changes.Last_Seq
-
-	// Update "fashion" doc to grant zegpold the role "hipster" and take it away from alice:
-	cacheWaiter.Add(1)
-	str := fmt.Sprintf(`{"user":"zegpold", "role":"role:hipster", "_rev":%q}`, fashionRevID)
-	RequireStatus(t, rt.Send(Request("PUT", "/db/fashion", str)), 201)
-
-	updatedRoleGrantSequence := rt.GetDocumentSequence("fashion")
-
-	// Check user access again:
-	alice, _ = a.GetUser("alice")
-	assert.Equal(t,
-
-		channels.TimedSet{
-			"!":     channels.NewVbSimpleSequence(0x1),
-			"alpha": channels.NewVbSimpleSequence(alice.Sequence()),
-		}, alice.InheritedChannels())
-
-	zegpold, _ = a.GetUser("zegpold")
-	assert.Equal(t,
-
-		channels.TimedSet{
-			"!":     channels.NewVbSimpleSequence(0x1),
-			"beta":  channels.NewVbSimpleSequence(zegpold.Sequence()),
-			"gamma": channels.NewVbSimpleSequence(updatedRoleGrantSequence),
-		}, zegpold.InheritedChannels())
-
-	// The complete _changes feed for zegpold contains docs g1 and b1:
-	cacheWaiter.Wait()
-	changes.Results = nil
-	response = rt.Send(RequestByUser("GET", "/db/_changes", "", "zegpold"))
-	log.Printf("3rd _changes looks like: %s", response.Body.Bytes())
-	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &changes))
-	log.Printf("changes: %+v", changes.Results)
-	require.Equal(t, len(changes.Results), 3)
-	assert.Equal(t, "_user/zegpold", changes.Results[0].ID)
-	assert.Equal(t, "b1", changes.Results[1].ID)
-	assert.Equal(t, "g1", changes.Results[2].ID)
-
-	// Changes feed with since=lastSeqPreGrant would ordinarily be empty, but zegpold got access to channel
-	// gamma after lastSeqPreGrant, so the pre-existing docs in that channel are included:
-	response = rt.Send(RequestByUser("GET", fmt.Sprintf("/db/_changes?since=%v", lastSeqPreGrant), "", "zegpold"))
-	log.Printf("4th _changes looks like: %s", response.Body.Bytes())
-	changes.Results = nil
-	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &changes))
-	require.Equal(t, 1, len(changes.Results))
-	assert.Equal(t, "g1", changes.Results[0].ID)
 }
 
 func TestAllDocsChannelsAfterChannelMove(t *testing.T) {
@@ -3818,225 +3452,6 @@ func TestUptimeStat(t *testing.T) {
 	assert.True(t, uptime1 < uptime2)
 }
 
-func TestRemovingUserXattr(t *testing.T) {
-	if base.UnitTestUrlIsWalrus() {
-		t.Skip("This test only works against Couchbase Server")
-	}
-
-	if !base.TestUseXattrs() {
-		t.Skip("This test only works with XATTRS enabled")
-	}
-
-	if !base.IsEnterpriseEdition() {
-		t.Skipf("test is EE only - user xattrs")
-	}
-
-	defer db.SuspendSequenceBatching()()
-
-	base.SetUpTestLogging(t, base.LevelDebug, base.KeyAll)
-
-	testCases := []struct {
-		name          string
-		autoImport    bool
-		importTrigger func(t *testing.T, rt *RestTester, docKey string)
-	}{
-		{
-			name:       "GET",
-			autoImport: false,
-			importTrigger: func(t *testing.T, rt *RestTester, docKey string) {
-				resp := rt.SendAdminRequest("GET", "/db/"+docKey, "")
-				RequireStatus(t, resp, http.StatusOK)
-			},
-		},
-		{
-			name:       "PUT",
-			autoImport: false,
-			importTrigger: func(t *testing.T, rt *RestTester, docKey string) {
-				resp := rt.SendAdminRequest("PUT", "/db/"+docKey, "{}")
-				RequireStatus(t, resp, http.StatusConflict)
-			},
-		},
-		{
-			name:       "Auto",
-			autoImport: true,
-			importTrigger: func(t *testing.T, rt *RestTester, docKey string) {
-				// No op
-			},
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			docKey := testCase.name
-			xattrKey := "myXattr"
-			channelName := "testChan"
-
-			// Sync function to set channel access to whatever xattr is
-			rt := NewRestTester(t, &RestTesterConfig{
-				DatabaseConfig: &DatabaseConfig{DbConfig: DbConfig{
-					AutoImport:   testCase.autoImport,
-					UserXattrKey: xattrKey,
-				}},
-				SyncFn: `
-			function (doc, oldDoc, meta){
-				if (meta.xattrs.myXattr !== undefined){
-					channel(meta.xattrs.myXattr);
-				}
-			}`,
-			})
-
-			defer rt.Close()
-
-			gocbBucket, ok := base.AsUserXattrStore(rt.Bucket())
-			if !ok {
-				t.Skip("Test requires Couchbase Bucket")
-			}
-
-			// Initial PUT
-			resp := rt.SendAdminRequest("PUT", "/db/"+docKey, `{}`)
-			RequireStatus(t, resp, http.StatusCreated)
-
-			// Add xattr
-			_, err := gocbBucket.WriteUserXattr(docKey, xattrKey, channelName)
-			assert.NoError(t, err)
-
-			// Trigger import
-			testCase.importTrigger(t, rt, docKey)
-			err = rt.WaitForCondition(func() bool {
-				return rt.GetDatabase().DbStats.SharedBucketImport().ImportCount.Value() == 1
-			})
-			assert.NoError(t, err)
-
-			// Get sync data for doc and ensure user xattr has been used correctly to set channel
-			var syncData db.SyncData
-			subdocStore, ok := base.AsSubdocXattrStore(rt.Bucket())
-			require.True(t, ok)
-			_, err = subdocStore.SubdocGetXattr(docKey, base.SyncXattrName, &syncData)
-			assert.NoError(t, err)
-
-			assert.Equal(t, []string{channelName}, syncData.Channels.KeySet())
-
-			// Delete user xattr
-			_, err = gocbBucket.DeleteUserXattr(docKey, xattrKey)
-			assert.NoError(t, err)
-
-			// Trigger import
-			testCase.importTrigger(t, rt, docKey)
-			err = rt.WaitForCondition(func() bool {
-				return rt.GetDatabase().DbStats.SharedBucketImport().ImportCount.Value() == 2
-			})
-			assert.NoError(t, err)
-
-			// Ensure old channel set with user xattr has been removed
-			var syncData2 db.SyncData
-			_, err = subdocStore.SubdocGetXattr(docKey, base.SyncXattrName, &syncData2)
-			assert.NoError(t, err)
-
-			assert.Equal(t, uint64(3), syncData2.Channels[channelName].Seq)
-		})
-	}
-}
-
-func TestUserXattrAvoidRevisionIDGeneration(t *testing.T) {
-	if base.UnitTestUrlIsWalrus() {
-		t.Skip("This test only works against Couchbase Server")
-	}
-
-	if !base.TestUseXattrs() {
-		t.Skip("This test only works with XATTRS enabled")
-	}
-
-	if !base.IsEnterpriseEdition() {
-		t.Skipf("test is EE only - user xattrs")
-	}
-
-	base.SetUpTestLogging(t, base.LevelDebug, base.KeyAll)
-
-	docKey := t.Name()
-	xattrKey := "myXattr"
-	channelName := "testChan"
-
-	// Sync function to set channel access to whatever xattr is
-	rt := NewRestTester(t, &RestTesterConfig{
-		DatabaseConfig: &DatabaseConfig{DbConfig: DbConfig{
-			AutoImport:   true,
-			UserXattrKey: xattrKey,
-		}},
-		SyncFn: `
-			function (doc, oldDoc, meta){
-				if (meta.xattrs.myXattr !== undefined){
-					channel(meta.xattrs.myXattr);
-					console.log(JSON.stringify(meta));
-				}
-			}`,
-	})
-
-	defer rt.Close()
-
-	userXattrStore, ok := base.AsUserXattrStore(rt.Bucket())
-	if !ok {
-		t.Skip("Test requires Couchbase Bucket")
-	}
-
-	subdocXattrStore, ok := base.AsSubdocXattrStore(rt.Bucket())
-	require.True(t, ok)
-
-	// Initial PUT
-	resp := rt.SendAdminRequest("PUT", "/db/"+docKey, `{}`)
-	RequireStatus(t, resp, http.StatusCreated)
-
-	require.NoError(t, rt.WaitForPendingChanges())
-
-	// Get current sync data
-	var syncData db.SyncData
-	_, err := subdocXattrStore.SubdocGetXattr(docKey, base.SyncXattrName, &syncData)
-	assert.NoError(t, err)
-
-	docRev, err := rt.GetDatabase().GetSingleDatabaseCollection().GetRevisionCacheForTest().Get(base.TestCtx(t), docKey, syncData.CurrentRev, true, false)
-	assert.NoError(t, err)
-	assert.Equal(t, 0, len(docRev.Channels.ToArray()))
-	assert.Equal(t, syncData.CurrentRev, docRev.RevID)
-
-	// Write xattr to trigger import of user xattr
-	_, err = userXattrStore.WriteUserXattr(docKey, xattrKey, channelName)
-	assert.NoError(t, err)
-
-	// Wait for import
-	err = rt.WaitForCondition(func() bool {
-		return rt.GetDatabase().DbStats.SharedBucketImport().ImportCount.Value() == 1
-	})
-	assert.NoError(t, err)
-
-	// Ensure import worked and sequence incremented but that sequence did not
-	var syncData2 db.SyncData
-	_, err = subdocXattrStore.SubdocGetXattr(docKey, base.SyncXattrName, &syncData2)
-	assert.NoError(t, err)
-
-	docRev2, err := rt.GetDatabase().GetSingleDatabaseCollection().GetRevisionCacheForTest().Get(base.TestCtx(t), docKey, syncData.CurrentRev, true, false)
-	assert.NoError(t, err)
-	assert.Equal(t, syncData2.CurrentRev, docRev2.RevID)
-
-	assert.Equal(t, syncData.CurrentRev, syncData2.CurrentRev)
-	assert.True(t, syncData2.Sequence > syncData.Sequence)
-	assert.Equal(t, []string{channelName}, syncData2.Channels.KeySet())
-	assert.Equal(t, syncData2.Channels.KeySet(), docRev2.Channels.ToArray())
-
-	err = rt.Bucket().Set(docKey, 0, nil, []byte(`{"update": "update"}`))
-	assert.NoError(t, err)
-
-	err = rt.WaitForCondition(func() bool {
-		log.Printf("Import count is: %v", rt.GetDatabase().DbStats.SharedBucketImport().ImportCount.Value())
-		return rt.GetDatabase().DbStats.SharedBucketImport().ImportCount.Value() == 2
-	})
-	assert.NoError(t, err)
-
-	var syncData3 db.SyncData
-	_, err = subdocXattrStore.SubdocGetXattr(docKey, base.SyncXattrName, &syncData2)
-	assert.NoError(t, err)
-
-	assert.NotEqual(t, syncData2.CurrentRev, syncData3.CurrentRev)
-}
-
 func TestDocumentChannelHistory(t *testing.T) {
 	defer db.SuspendSequenceBatching()()
 
@@ -4151,10 +3566,6 @@ func TestChannelHistoryLegacyDoc(t *testing.T) {
 		End:   2,
 	})
 	assert.Len(t, syncData.ChannelSetHistory, 0)
-}
-
-type UserRolesTemp struct {
-	Roles map[string][]string `json:"roles"`
 }
 
 type ChannelsTemp struct {
