+++ rest/user_api_test.go
@@ -9,9 +9,16 @@
 package rest
 
 import (
+	"bufio"
+	"bytes"
 	"encoding/json"
 	"fmt"
+	"io"
+	"log"
+	"net/http"
+	"sync"
 	"testing"
+	"time"
 
 	"github.com/couchbase/sync_gateway/auth"
 	"github.com/couchbase/sync_gateway/base"
@@ -146,60 +153,6 @@ func TestUsersAPIDetailsWithoutGuest(t *testing.T) {
 
 	validateUsersNameOnlyFalse(t, rt)
 
-}
-
-// validateUsersNameOnlyFalse validates query results with paramNameOnly=false.  Includes
-// users with/without email, enabled/disabled, various limits
-func validateUsersNameOnlyFalse(t *testing.T, rt *RestTester) {
-	// Validate the zero user case
-	var responseUsers []auth.PrincipalConfig
-	response := rt.SendAdminRequest("GET", "/db/_user/?"+paramNameOnly+"=false", "")
-	RequireStatus(t, response, 200)
-	err := json.Unmarshal(response.Body.Bytes(), &responseUsers)
-	assert.NoError(t, err)
-	assert.Equal(t, 0, len(responseUsers))
-
-	// Test for user counts going from 1 to a few multiples of QueryPaginationLimit to check boundary conditions.
-	for i := 1; i <= 12; i++ {
-		userName := fmt.Sprintf("user%d", i)
-		disabled := "false"
-		if userName == "user3" || userName == "user8" {
-			disabled = "true"
-		}
-		// don't set email in some users
-		var response *TestResponse
-		if userName == "user5" {
-			response = rt.SendAdminRequest("PUT", "/db/_user/"+userName, `{"password":"letmein", "disabled":`+disabled+`, "admin_channels":["foo", "bar"]}`)
-		} else {
-			response = rt.SendAdminRequest("PUT", "/db/_user/"+userName, `{"password":"letmein", "email": "`+userName+`@foo.com", "disabled":`+disabled+`, "admin_channels":["foo", "bar"]}`)
-		}
-		RequireStatus(t, response, 201)
-
-		// check user count
-		response = rt.SendAdminRequest("GET", "/db/_user/?"+paramNameOnly+"=false", "")
-		RequireStatus(t, response, 200)
-		err := json.Unmarshal(response.Body.Bytes(), &responseUsers)
-		assert.NoError(t, err)
-		assert.Equal(t, i, len(responseUsers))
-
-		// Check property values, and validate no duplicate users returned in response
-		userMap := make(map[string]interface{})
-		for _, principal := range responseUsers {
-			require.NotNil(t, principal.Name)
-			if *principal.Name != "user5" {
-				require.NotNil(t, principal.Email)
-				assert.Equal(t, *principal.Name+"@foo.com", *principal.Email)
-			}
-			if *principal.Name == "user3" || *principal.Name == "user8" {
-				assert.Equal(t, true, *principal.Disabled)
-			} else {
-				assert.Equal(t, false, *principal.Disabled)
-			}
-			_, ok := userMap[*principal.Name]
-			assert.False(t, ok)
-			userMap[*principal.Name] = struct{}{}
-		}
-	}
 }
 
 // TestUsersAPIDetailsWithLimit tests users endpoint with name_only=false and limit
@@ -413,4 +366,990 @@ func TestUserAPI(t *testing.T) {
 	// DELETE the user, double escape username, will succeed
 	RequireStatus(t, rt.SendAdminRequest("DELETE", "/db/_user/0%257C%4059", ""), 200)
 
+}
+
+func TestGuestUser(t *testing.T) {
+
+	guestUserEndpoint := fmt.Sprintf("/db/_user/%s", base.GuestUsername)
+
+	rt := NewRestTester(t, nil)
+	defer rt.Close()
+
+	response := rt.SendAdminRequest(http.MethodGet, guestUserEndpoint, "")
+	RequireStatus(t, response, http.StatusOK)
+	var body db.Body
+	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &body))
+	assert.Equal(t, base.GuestUsername, body["name"])
+	// This ain't no admin-party, this ain't no nightclub, this ain't no fooling around:
+	assert.Nil(t, body["admin_channels"])
+
+	// Disable the guest user:
+	response = rt.SendAdminRequest(http.MethodPut, guestUserEndpoint, `{"disabled":true}`)
+	RequireStatus(t, response, http.StatusOK)
+
+	// Get guest user and verify it is now disabled:
+	response = rt.SendAdminRequest(http.MethodGet, guestUserEndpoint, "")
+	RequireStatus(t, response, http.StatusOK)
+	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &body))
+	assert.Equal(t, base.GuestUsername, body["name"])
+	assert.True(t, body["disabled"].(bool))
+
+	// Check that the actual User object is correct:
+	ctx := rt.Context()
+	user, _ := rt.ServerContext().Database(ctx, "db").Authenticator(ctx).GetUser("")
+	assert.Empty(t, user.Name())
+	assert.Nil(t, user.ExplicitChannels())
+	assert.True(t, user.Disabled())
+
+	// We can't delete the guest user, but we should get a reasonable error back.
+	response = rt.SendAdminRequest(http.MethodDelete, guestUserEndpoint, "")
+	RequireStatus(t, response, http.StatusMethodNotAllowed)
+}
+
+func TestUserAndRoleResponseContentType(t *testing.T) {
+	rt := NewRestTester(t, nil)
+	defer rt.Close()
+
+	// Create a user 'christopher' through PUT request with empty request body.
+	var responseBody db.Body
+	body := `{"email":"christopher@couchbase.com","password":"cGFzc3dvcmQ=","admin_channels":["foo", "bar"]}`
+	response := rt.SendAdminRequest(http.MethodPut, "/db/_user/christopher", "")
+	assert.Equal(t, http.StatusBadRequest, response.Code)
+	assert.Equal(t, "application/json", response.Header().Get("Content-Type"))
+	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &responseBody))
+
+	// Create a user 'charles' through POST request with empty request body.
+	body = `{"email":"charles@couchbase.com","password":"cGFzc3dvcmQ=","admin_channels":["foo", "bar"]}`
+	response = rt.SendAdminRequest(http.MethodPost, "/db/_user/charles", "")
+	assert.Equal(t, http.StatusMethodNotAllowed, response.Code)
+	assert.Equal(t, "application/json", response.Header().Get("Content-Type"))
+	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &responseBody))
+
+	// Create a user 'alice' through PUT request.
+	body = `{"email":"alice@couchbase.com","password":"cGFzc3dvcmQ=","admin_channels":["foo", "bar"]}`
+	response = rt.SendAdminRequest(http.MethodPut, "/db/_user/alice", body)
+	assert.Equal(t, http.StatusCreated, response.Code)
+	assert.Empty(t, response.Header().Get("Content-Type"))
+
+	// Create another user 'bob' through POST request.
+	body = `{"name":"bob","email":"bob@couchbase.com","password":"cGFzc3dvcmQ=","admin_channels":["foo", "bar"]}`
+	response = rt.SendAdminRequest(http.MethodPost, "/db/_user/", body)
+	assert.Equal(t, http.StatusCreated, response.Code)
+	assert.Empty(t, response.Header().Get("Content-Type"))
+
+	// Get the user details of user 'alice' through GET request.
+	response = rt.SendAdminRequest(http.MethodGet, "/db/_user/alice", "")
+	assert.Equal(t, http.StatusOK, response.Code)
+	assert.Equal(t, "application/json", response.Header().Get("Content-Type"))
+	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &responseBody))
+
+	// Get the list of users through GET request.
+	var users []string
+	response = rt.SendAdminRequest(http.MethodGet, "/db/_user/", "")
+	assert.Equal(t, http.StatusOK, response.Code)
+	assert.Equal(t, "application/json", response.Header().Get("Content-Type"))
+	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &users))
+	assert.Subset(t, []string{"alice", "bob"}, users)
+
+	// Check whether the /db/_user/bob resource exist on the server.
+	response = rt.SendAdminRequest(http.MethodHead, "/db/_user/bob", "")
+	assert.Equal(t, http.StatusOK, response.Code)
+	assert.Empty(t, response.Header().Get("Content-Type"))
+
+	// Get the list of users through HEAD request.
+	response = rt.SendAdminRequest(http.MethodHead, "/db/_user/", "")
+	assert.Equal(t, http.StatusOK, response.Code)
+	assert.Empty(t, response.Header().Get("Content-Type"))
+
+	// Delete user 'alice'
+	response = rt.SendAdminRequest(http.MethodDelete, "/db/_user/alice", "")
+	assert.Equal(t, http.StatusOK, response.Code)
+	assert.Empty(t, response.Header().Get("Content-Type"))
+
+	// Delete GUEST user instead of disabling.
+	response = rt.SendAdminRequest(http.MethodDelete, "/db/_user/GUEST", "")
+	assert.Equal(t, http.StatusMethodNotAllowed, response.Code)
+	assert.Equal(t, "application/json", response.Header().Get("Content-Type"))
+
+	// Delete user 'eve' who doesn't exists at this point of time.
+	response = rt.SendAdminRequest(http.MethodDelete, "/db/_user/eve", "")
+	assert.Equal(t, http.StatusNotFound, response.Code)
+	assert.Equal(t, "application/json", response.Header().Get("Content-Type"))
+
+	// Create a new user and save to database to create user session.
+	authenticator := auth.NewAuthenticator(rt.Bucket(), nil, auth.DefaultAuthenticatorOptions())
+	user, err := authenticator.NewUser("eve", "cGFzc3dvcmQ=", channels.BaseSetOf(t, "*"))
+	assert.NoError(t, err, "Couldn't create new user")
+	assert.NoError(t, authenticator.Save(user), "Couldn't save new user")
+
+	// Create user session to check delete session request.
+	response = rt.SendAdminRequest(http.MethodPost, "/db/_session", `{"name":"eve"}`)
+	assert.Equal(t, http.StatusOK, response.Code)
+	assert.Equal(t, "application/json", response.Header().Get("Content-Type"))
+	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &responseBody))
+	sessionId, _ := responseBody["session_id"].(string)
+	require.NotEmpty(t, sessionId, "Couldn't parse sessionID from response body")
+
+	// Delete user session using /db/_user/eve/_session/{sessionId}.
+	response = rt.SendAdminRequest(http.MethodDelete, "/db/_user/eve/_session/"+sessionId, "")
+	assert.Equal(t, http.StatusOK, response.Code)
+	assert.Empty(t, response.Header().Get("Content-Type"))
+
+	// Create user session to check delete session request.
+	response = rt.SendAdminRequest(http.MethodPost, "/db/_session", `{"name":"eve"}`)
+	assert.Equal(t, http.StatusOK, response.Code)
+	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &responseBody))
+
+	// Delete user session using /db/_user/eve/_session request.
+	response = rt.SendAdminRequest(http.MethodDelete, "/db/_user/eve/_session", "")
+	assert.Equal(t, http.StatusOK, response.Code)
+	assert.Empty(t, response.Header().Get("Content-Type"))
+
+	// Create a role 'developer' through POST request
+	body = `{"name":"developer","admin_channels":["channel1", "channel2"]}`
+	response = rt.SendAdminRequest(http.MethodPost, "/db/_role/", body)
+	assert.Equal(t, http.StatusCreated, response.Code)
+	assert.Empty(t, response.Header().Get("Content-Type"))
+
+	// Create another role 'coder' through PUT request.
+	body = `{"admin_channels":["channel3", "channel4"]}`
+	response = rt.SendAdminRequest(http.MethodPut, "/db/_role/coder", body)
+	assert.Equal(t, http.StatusCreated, response.Code)
+	assert.Empty(t, response.Header().Get("Content-Type"))
+
+	// Check whether the /db/_role/ resource exist on the server.
+	response = rt.SendAdminRequest(http.MethodHead, "/db/_role/", "")
+	assert.Equal(t, http.StatusOK, response.Code)
+	assert.Empty(t, response.Header().Get("Content-Type"))
+
+	// Get the created roles through GET request.
+	var roles []string
+	response = rt.SendAdminRequest(http.MethodGet, "/db/_role/", "")
+	assert.Equal(t, http.StatusOK, response.Code)
+	assert.Equal(t, "application/json", response.Header().Get("Content-Type"))
+	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &roles))
+	assert.Subset(t, []string{"coder", "developer"}, roles)
+
+	// Delete role 'coder' from database.
+	response = rt.SendAdminRequest(http.MethodDelete, "/db/_role/coder", "")
+	assert.Equal(t, http.StatusOK, response.Code)
+	assert.Empty(t, response.Header().Get("Content-Type"))
+
+	// Delete role who doesn't exist.
+	response = rt.SendAdminRequest(http.MethodDelete, "/db/_role/programmer", "")
+	assert.Equal(t, http.StatusNotFound, response.Code)
+	assert.Equal(t, "application/json", response.Header().Get("Content-Type"))
+	require.NoError(t, base.JSONUnmarshal(response.Body.Bytes(), &responseBody))
+}
+
+func TestUserXattrsRawGet(t *testing.T) {
+	if !base.TestUseXattrs() {
+		t.Skip("Test requires xattrs to be enabled")
+	}
+
+	if !base.IsEnterpriseEdition() {
+		t.Skipf("test is EE only - user xattrs")
+	}
+
+	docKey := t.Name()
+	xattrKey := "xattrKey"
+
+	rt := NewRestTester(t, &RestTesterConfig{
+		DatabaseConfig: &DatabaseConfig{
+			DbConfig: DbConfig{
+				AutoImport:   true,
+				UserXattrKey: xattrKey,
+			},
+		},
+	})
+	defer rt.Close()
+
+	userXattrStore, ok := base.AsUserXattrStore(rt.Bucket())
+	if !ok {
+		t.Skip("Test requires Couchbase Bucket")
+	}
+
+	resp := rt.SendAdminRequest("PUT", "/db/"+docKey, "{}")
+	RequireStatus(t, resp, http.StatusCreated)
+	require.NoError(t, rt.WaitForPendingChanges())
+
+	_, err := userXattrStore.WriteUserXattr(docKey, xattrKey, "val")
+	assert.NoError(t, err)
+
+	err = rt.WaitForCondition(func() bool {
+		return rt.GetDatabase().DbStats.SharedBucketImportStats.ImportCount.Value() == 1
+	})
+
+	resp = rt.SendAdminRequest("GET", "/db/_raw/"+docKey, "")
+	RequireStatus(t, resp, http.StatusOK)
+
+	var RawReturn struct {
+		Meta struct {
+			Xattrs map[string]interface{} `json:"xattrs"`
+		} `json:"_meta"`
+	}
+
+	err = json.Unmarshal(resp.BodyBytes(), &RawReturn)
+
+	assert.Equal(t, "val", RawReturn.Meta.Xattrs[xattrKey])
+}
+
+func TestObtainUserChannelsForDeletedRoleCasFail(t *testing.T) {
+	if !base.UnitTestUrlIsWalrus() {
+		t.Skip("Skip LeakyBucket test when running in integration")
+	}
+
+	testCases := []struct {
+		Name      string
+		RunBefore bool
+	}{
+		{
+			"Delete On GetUser",
+			true,
+		},
+		{
+			"Delete On InheritedChannels",
+			false,
+		},
+	}
+
+	for _, testCase := range testCases {
+		t.Run(testCase.Name, func(t *testing.T) {
+			rt := NewRestTester(t, &RestTesterConfig{
+				SyncFn: `
+			function(doc, oldDoc){
+				if (doc._id === 'roleChannels'){
+					access('role:role', doc.channels)
+				}
+				if (doc._id === 'userRoles'){
+					role('user', doc.roles)
+				}
+			}
+		`,
+			})
+			defer rt.Close()
+
+			// Create role
+			resp := rt.SendAdminRequest("PUT", "/db/_role/role", `{"admin_channels":["channel"]}`)
+			RequireStatus(t, resp, http.StatusCreated)
+
+			// Create user
+			resp = rt.SendAdminRequest("PUT", "/db/_user/user", `{"password": "pass"}`)
+			RequireStatus(t, resp, http.StatusCreated)
+
+			// Add channel to role
+			resp = rt.SendAdminRequest("PUT", "/db/roleChannels", `{"channels": "inherit"}`)
+			RequireStatus(t, resp, http.StatusCreated)
+
+			// Add role to user
+			resp = rt.SendAdminRequest("PUT", "/db/userRoles", `{"roles": "role:role"}`)
+			RequireStatus(t, resp, http.StatusCreated)
+
+			leakyBucket, ok := base.AsLeakyBucket(rt.TestBucket)
+			require.True(t, ok)
+
+			triggerCallback := false
+			leakyBucket.SetUpdateCallback(func(key string) {
+				if triggerCallback {
+					triggerCallback = false
+					resp = rt.SendAdminRequest("DELETE", "/db/_role/role", ``)
+					RequireStatus(t, resp, http.StatusOK)
+				}
+			})
+
+			if testCase.RunBefore {
+				triggerCallback = true
+			}
+
+			authenticator := rt.GetDatabase().Authenticator(base.TestCtx(t))
+			user, err := authenticator.GetUser("user")
+			assert.NoError(t, err)
+
+			if !testCase.RunBefore {
+				triggerCallback = true
+			}
+
+			assert.Equal(t, []string{"!"}, user.InheritedChannels().AllKeys())
+
+			// Ensure callback ran
+			assert.False(t, triggerCallback)
+		})
+
+	}
+}
+
+func TestUserPasswordValidation(t *testing.T) {
+
+	// PUT a user
+	rt := NewRestTester(t, nil)
+	defer rt.Close()
+
+	response := rt.SendAdminRequest("PUT", "/db/_user/snej", `{"email":"jens@couchbase.com", "password":"letmein", "admin_channels":["foo", "bar"]}`)
+	RequireStatus(t, response, 201)
+
+	// PUT a user without a password, should fail
+	response = rt.SendAdminRequest("PUT", "/db/_user/ajresnopassword", `{"email":"ajres@couchbase.com", "admin_channels":["foo", "bar"]}`)
+	RequireStatus(t, response, 400)
+
+	// POST a user without a password, should fail
+	response = rt.SendAdminRequest("POST", "/db/_user/", `{"name":"ajresnopassword", "email":"ajres@couchbase.com", "admin_channels":["foo", "bar"]}`)
+	RequireStatus(t, response, 400)
+
+	// PUT a user with a two character password, should fail
+	response = rt.SendAdminRequest("PUT", "/db/_user/ajresnopassword", `{"email":"ajres@couchbase.com", "password":"in", "admin_channels":["foo", "bar"]}`)
+	RequireStatus(t, response, 400)
+
+	// POST a user with a two character password, should fail
+	response = rt.SendAdminRequest("POST", "/db/_user/", `{"name":"ajresnopassword", "email":"ajres@couchbase.com", "password":"an", "admin_channels":["foo", "bar"]}`)
+	RequireStatus(t, response, 400)
+
+	// PUT a user with a zero character password, should fail
+	response = rt.SendAdminRequest("PUT", "/db/_user/ajresnopassword", `{"email":"ajres@couchbase.com", "password":"", "admin_channels":["foo", "bar"]}`)
+	RequireStatus(t, response, 400)
+
+	// POST a user with a zero character password, should fail
+	response = rt.SendAdminRequest("POST", "/db/_user/", `{"name":"ajresnopassword", "email":"ajres@couchbase.com", "password":"", "admin_channels":["foo", "bar"]}`)
+	RequireStatus(t, response, 400)
+
+	// PUT update a user with a two character password, should fail
+	response = rt.SendAdminRequest("PUT", "/db/_user/snej", `{"password":"an"}`)
+	RequireStatus(t, response, 400)
+
+	// PUT update a user with a one character password, should fail
+	response = rt.SendAdminRequest("PUT", "/db/_user/snej", `{"password":"a"}`)
+	RequireStatus(t, response, 400)
+
+	// PUT update a user with a zero character password, should fail
+	response = rt.SendAdminRequest("PUT", "/db/_user/snej", `{"password":""}`)
+	RequireStatus(t, response, 400)
+
+	// PUT update a user with a three character password, should succeed
+	response = rt.SendAdminRequest("PUT", "/db/_user/snej", `{"password":"abc"}`)
+	RequireStatus(t, response, 200)
+}
+func TestUserAllowEmptyPassword(t *testing.T) {
+
+	// PUT a user
+	rt := NewRestTester(t, &RestTesterConfig{DatabaseConfig: &DatabaseConfig{DbConfig: DbConfig{AllowEmptyPassword: base.BoolPtr(true)}}})
+	defer rt.Close()
+
+	response := rt.SendAdminRequest("PUT", "/db/_user/snej", `{"email":"jens@couchbase.com", "password":"letmein", "admin_channels":["foo", "bar"]}`)
+	RequireStatus(t, response, 201)
+
+	// PUT a user without a password, should succeed
+	response = rt.SendAdminRequest("PUT", "/db/_user/nopassword1", `{"email":"ajres@couchbase.com", "admin_channels":["foo", "bar"]}`)
+	RequireStatus(t, response, 201)
+
+	// POST a user without a password, should succeed
+	response = rt.SendAdminRequest("POST", "/db/_user/", `{"name":"nopassword2", "email":"ajres@couchbase.com", "admin_channels":["foo", "bar"]}`)
+	RequireStatus(t, response, 201)
+
+	// PUT a user with a two character password, should fail
+	response = rt.SendAdminRequest("PUT", "/db/_user/nopassword3", `{"email":"ajres@couchbase.com", "password":"in", "admin_channels":["foo", "bar"]}`)
+	RequireStatus(t, response, 400)
+
+	// POST a user with a two character password, should fail
+	response = rt.SendAdminRequest("POST", "/db/_user/", `{"name":"nopassword4", "email":"ajres@couchbase.com", "password":"an", "admin_channels":["foo", "bar"]}`)
+	RequireStatus(t, response, 400)
+
+	// PUT a user with a zero character password, should succeed
+	response = rt.SendAdminRequest("PUT", "/db/_user/nopassword5", `{"email":"ajres@couchbase.com", "password":"", "admin_channels":["foo", "bar"]}`)
+	RequireStatus(t, response, 201)
+
+	// POST a user with a zero character password, should succeed
+	response = rt.SendAdminRequest("POST", "/db/_user/", `{"name":"nopassword6", "email":"ajres@couchbase.com", "password":"", "admin_channels":["foo", "bar"]}`)
+	RequireStatus(t, response, 201)
+
+	// PUT update a user with a two character password, should fail
+	response = rt.SendAdminRequest("PUT", "/db/_user/snej", `{"password":"an"}`)
+	RequireStatus(t, response, 400)
+
+	// PUT update a user with a one character password, should fail
+	response = rt.SendAdminRequest("PUT", "/db/_user/snej", `{"password":"a"}`)
+	RequireStatus(t, response, 400)
+
+	// PUT update a user with a zero character password, should succeed
+	response = rt.SendAdminRequest("PUT", "/db/_user/snej", `{"password":""}`)
+	RequireStatus(t, response, 200)
+
+	// PUT update a user with a three character password, should succeed
+	response = rt.SendAdminRequest("PUT", "/db/_user/snej", `{"password":"abc"}`)
+	RequireStatus(t, response, 200)
+}
+
+func TestPrincipalForbidUpdatingChannels(t *testing.T) {
+	rt := NewRestTester(t, &RestTesterConfig{})
+	defer rt.Close()
+
+	// Users
+	// PUT admin_channels
+	response := rt.SendAdminRequest("PUT", "/db/_user/snej", `{"email":"jens@couchbase.com", "password":"letmein", "admin_channels":["foo", "bar"]}`)
+	RequireStatus(t, response, 201)
+
+	// PUT all_channels - should fail
+	response = rt.SendAdminRequest("PUT", "/db/_user/snej", `{"email":"jens@couchbase.com", "all_channels":["baz"]}`)
+	RequireStatus(t, response, 400)
+
+	// PUT admin_roles
+	response = rt.SendAdminRequest("PUT", "/db/_user/snej", `{"email":"jens@couchbase.com", "password":"letmein", "admin_roles":["foo", "bar"]}`)
+	RequireStatus(t, response, 200)
+
+	// PUT roles - should fail
+	response = rt.SendAdminRequest("PUT", "/db/_user/snej", `{"email":"jens@couchbase.com", "roles":["baz"]}`)
+	RequireStatus(t, response, 400)
+
+	// Roles
+	// PUT admin_channels
+	response = rt.SendAdminRequest("PUT", "/db/_role/test", `{"admin_channels":["foo", "bar"]}`)
+	RequireStatus(t, response, 201)
+
+	// PUT all_channels - should fail
+	response = rt.SendAdminRequest("PUT", "/db/_role/test", `{"all_channels":["baz"]}`)
+	RequireStatus(t, response, 400)
+}
+
+// Test user access grant while that user has an active changes feed.  (see issue #880)
+func TestUserAccessRace(t *testing.T) {
+
+	base.LongRunningTest(t)
+
+	// This test only runs against Walrus due to known sporadic failures.
+	// See https://github.com/couchbase/sync_gateway/issues/3006
+	if !base.UnitTestUrlIsWalrus() {
+		t.Skip("Skip this test under integration testing")
+	}
+
+	syncFunction := `
+function(doc, oldDoc) {
+  if (doc.type == "list") {
+    channel("list-"+doc._id);
+  } else if (doc.type == "profile") {
+    channel("profiles");
+    var user = doc._id.substring(doc._id.indexOf(":")+1);
+    if (user !== doc.user_id) {
+      throw({forbidden : "profile user_id must match docid"})
+    }
+    requireUser(user);
+    access(user, "profiles");
+    channel('profile-'+user);
+  } else if (doc.type == "Want") {
+    var parts = doc._id.split("-");
+    var user = parts[1];
+    var i = parts[2];
+    requireUser(user);
+    channel('profile-'+user);
+    access(user, 'list-'+i);
+  }
+}
+
+`
+	rtConfig := RestTesterConfig{SyncFn: syncFunction}
+	var rt = NewRestTester(t, &rtConfig)
+	defer rt.Close()
+
+	response := rt.SendAdminRequest("PUT", "/db/_user/bernard", `{"name":"bernard", "password":"letmein", "admin_channels":["profile-bernard"]}`)
+	RequireStatus(t, response, 201)
+
+	// Try to force channel initialisation for user bernard
+	response = rt.SendAdminRequest("GET", "/db/_user/bernard", "")
+	RequireStatus(t, response, 200)
+
+	// Create list docs
+	input := `{"docs": [`
+
+	for i := 1; i <= 100; i++ {
+		if i > 1 {
+			input = input + `,`
+		}
+		docId := fmt.Sprintf("l_%d", i)
+		input = input + fmt.Sprintf(`{"_id":"%s", "type":"list"}`, docId)
+	}
+	input = input + `]}`
+	response = rt.SendAdminRequest("POST", "/db/_bulk_docs", input)
+
+	// Start changes feed
+	var wg sync.WaitGroup
+
+	// Init the public handler, to avoid data race initializing in the two usages below.
+	_ = rt.SendRequest("GET", "/db", "")
+
+	wg.Add(1)
+
+	numExpectedChanges := 201
+
+	go func() {
+		defer wg.Done()
+
+		since := ""
+
+		maxTries := 10
+		numTries := 0
+
+		changesAccumulated := []db.ChangeEntry{}
+
+		for {
+
+			// Timeout allows us to read continuous changes after processing is complete.  Needs to be long enough to
+			// ensure it doesn't terminate before the first change is sent.
+			log.Printf("Invoking _changes?feed=continuous&since=%s&timeout=2000", since)
+			changesResponse := rt.Send(RequestByUser("GET", fmt.Sprintf("/db/_changes?feed=continuous&since=%s&timeout=2000", since), "", "bernard"))
+
+			changes, err := readContinuousChanges(changesResponse)
+			assert.NoError(t, err)
+
+			changesAccumulated = append(changesAccumulated, changes...)
+
+			if len(changesAccumulated) >= numExpectedChanges {
+				log.Printf("Got numExpectedChanges (%d).  Done", numExpectedChanges)
+				break
+			} else {
+				log.Printf("Only received %d out of %d expected changes.  Attempt %d / %d.", len(changesAccumulated), numExpectedChanges, numTries, maxTries)
+			}
+
+			// Advance the since value if we got any changes
+			if len(changes) > 0 {
+				since = changes[len(changes)-1].Seq.String()
+				log.Printf("Setting since value to: %s.", since)
+			}
+
+			numTries++
+			if numTries > maxTries {
+				t.Errorf("Giving up trying to receive %d changes.  Only received %d", numExpectedChanges, len(changesAccumulated))
+				return
+			}
+
+		}
+
+	}()
+
+	// Make bulk docs calls, 100 docs each, all triggering access grants to the list docs.
+	for j := 0; j < 1; j++ {
+
+		input := `{"docs": [`
+		for i := 1; i <= 100; i++ {
+			if i > 1 {
+				input = input + `,`
+			}
+			k := j*100 + i
+			docId := fmt.Sprintf("Want-bernard-l_%d", k)
+			input = input + fmt.Sprintf(`{"_id":"%s", "type":"Want", "owner":"bernard"}`, docId)
+		}
+		input = input + `]}`
+
+		log.Printf("Sending 2nd round of _bulk_docs")
+		response = rt.Send(RequestByUser("POST", "/db/_bulk_docs", input, "bernard"))
+		log.Printf("Sent 2nd round of _bulk_docs")
+
+	}
+
+	// wait for changes feed to complete (time out)
+	wg.Wait()
+}
+
+// Test user delete while that user has an active changes feed (see issue 809)
+func TestUserDeleteDuringChangesWithAccess(t *testing.T) {
+
+	base.SetUpTestLogging(t, base.LevelInfo, base.KeyChanges, base.KeyCache, base.KeyHTTP)
+
+	rtConfig := RestTesterConfig{SyncFn: `function(doc) {channel(doc.channel); if(doc.type == "setaccess") { access(doc.owner, doc.channel);}}`}
+	rt := NewRestTester(t, &rtConfig)
+	defer rt.Close()
+
+	response := rt.SendAdminRequest("PUT", "/db/_user/bernard", `{"name":"bernard", "password":"letmein", "admin_channels":["foo"]}`)
+	RequireStatus(t, response, 201)
+
+	var wg sync.WaitGroup
+	wg.Add(1)
+	go func() {
+		defer wg.Done()
+		changesResponse := rt.Send(RequestByUser("GET", "/db/_changes?feed=continuous&since=0&timeout=3000", "", "bernard"))
+		// When testing single threaded, this reproduces the issue described in #809.
+		// When testing multithreaded (-cpu 4 -race), there are three (valid) possibilities"
+		// 1. The DELETE gets processed before the _changes auth completes: this will return 401
+		// 2. The _changes request gets processed before the DELETE: the changes response will be closed when the user is deleted
+		// 3. The DELETE is processed after the _changes auth completes, but before the MultiChangesFeed is instantiated.  The
+		//  changes feed doesn't have a trigger to attempt a reload of the user in this scenario, so will continue until disconnected
+		//  by the client.  This should be fixed more generally (to terminate all active user sessions when the user is deleted, not just
+		//  changes feeds) but that enhancement is too high risk to introduce at this time.  The timeout on changes will terminate the unit
+		//  test.
+		if changesResponse.Code == 401 {
+			// case 1 - ok
+		} else {
+			// case 2 - ensure no error processing the changes response.  The number of entries may vary, depending
+			// on whether the changes loop performed an additional iteration before catching the deleted user.
+			_, err := readContinuousChanges(changesResponse)
+			assert.NoError(t, err)
+		}
+	}()
+
+	// TODO: sleep required to ensure the changes feed iteration starts before the delete gets processed.
+	time.Sleep(500 * time.Millisecond)
+	rt.SendAdminRequest("PUT", "/db/bernard_doc1", `{"type":"setaccess", "owner":"bernard","channel":"foo"}`)
+	rt.SendAdminRequest("DELETE", "/db/_user/bernard", "")
+	rt.SendAdminRequest("PUT", "/db/manny_doc1", `{"type":"setaccess", "owner":"manny","channel":"bar"}`)
+	rt.SendAdminRequest("PUT", "/db/bernard_doc2", `{"type":"general", "channel":"foo"}`)
+
+	// case 3
+	for i := 0; i <= 5; i++ {
+		docId := fmt.Sprintf("/db/bernard_doc%d", i+3)
+		response = rt.SendAdminRequest("PUT", docId, `{"type":"setaccess", "owner":"bernard", "channel":"foo"}`)
+	}
+
+	wg.Wait()
+}
+
+// Helper functions
+
+// validateUsersNameOnlyFalse validates query results with paramNameOnly=false.  Includes
+// users with/without email, enabled/disabled, various limits
+func validateUsersNameOnlyFalse(t *testing.T, rt *RestTester) {
+	// Validate the zero user case
+	var responseUsers []auth.PrincipalConfig
+	response := rt.SendAdminRequest("GET", "/db/_user/?"+paramNameOnly+"=false", "")
+	RequireStatus(t, response, 200)
+	err := json.Unmarshal(response.Body.Bytes(), &responseUsers)
+	assert.NoError(t, err)
+	assert.Equal(t, 0, len(responseUsers))
+
+	// Test for user counts going from 1 to a few multiples of QueryPaginationLimit to check boundary conditions.
+	for i := 1; i <= 12; i++ {
+		userName := fmt.Sprintf("user%d", i)
+		disabled := "false"
+		if userName == "user3" || userName == "user8" {
+			disabled = "true"
+		}
+		// don't set email in some users
+		var response *TestResponse
+		if userName == "user5" {
+			response = rt.SendAdminRequest("PUT", "/db/_user/"+userName, `{"password":"letmein", "disabled":`+disabled+`, "admin_channels":["foo", "bar"]}`)
+		} else {
+			response = rt.SendAdminRequest("PUT", "/db/_user/"+userName, `{"password":"letmein", "email": "`+userName+`@foo.com", "disabled":`+disabled+`, "admin_channels":["foo", "bar"]}`)
+		}
+		RequireStatus(t, response, 201)
+
+		// check user count
+		response = rt.SendAdminRequest("GET", "/db/_user/?"+paramNameOnly+"=false", "")
+		RequireStatus(t, response, 200)
+		err := json.Unmarshal(response.Body.Bytes(), &responseUsers)
+		assert.NoError(t, err)
+		assert.Equal(t, i, len(responseUsers))
+
+		// Check property values, and validate no duplicate users returned in response
+		userMap := make(map[string]interface{})
+		for _, principal := range responseUsers {
+			require.NotNil(t, principal.Name)
+			if *principal.Name != "user5" {
+				require.NotNil(t, principal.Email)
+				assert.Equal(t, *principal.Name+"@foo.com", *principal.Email)
+			}
+			if *principal.Name == "user3" || *principal.Name == "user8" {
+				assert.Equal(t, true, *principal.Disabled)
+			} else {
+				assert.Equal(t, false, *principal.Disabled)
+			}
+			_, ok := userMap[*principal.Name]
+			assert.False(t, ok)
+			userMap[*principal.Name] = struct{}{}
+		}
+	}
+}
+
+// Reads continuous changes feed response into slice of ChangeEntry
+func readContinuousChanges(response *TestResponse) ([]db.ChangeEntry, error) {
+	var change db.ChangeEntry
+	changes := make([]db.ChangeEntry, 0)
+	reader := bufio.NewReader(response.Body)
+	for {
+		entry, readError := reader.ReadBytes('\n')
+		if readError == io.EOF {
+			// done
+			break
+		}
+		if readError != nil {
+			// unexpected read error
+			return changes, readError
+		}
+		entry = bytes.TrimSpace(entry)
+		if len(entry) > 0 {
+			err := base.JSONUnmarshal(entry, &change)
+			if err != nil {
+				return changes, err
+			}
+			changes = append(changes, change)
+			log.Printf("Got change ==> %v", change)
+		}
+
+	}
+	return changes, nil
+}
+func TestFunkyUsernames(t *testing.T) {
+	cases := []struct {
+		Name     string
+		UserName string
+	}{
+		{
+			Name:     "hashes",
+			UserName: "foo#bar",
+		},
+		{
+			Name:     "question mark",
+			UserName: "foo?bar",
+		},
+		{
+			Name:     "dollars",
+			UserName: "$foo$bar",
+		},
+		{
+			Name:     "underscore prefix",
+			UserName: "_sync-foobar",
+		},
+	}
+	for _, tc := range cases {
+		t.Run(tc.Name, func(t *testing.T) {
+			require.Truef(t, auth.IsValidPrincipalName(tc.UserName), "expected '%s' to be accepted", tc.UserName)
+			rt := NewRestTester(t, nil)
+			defer rt.Close()
+
+			ctx := rt.Context()
+			a := rt.ServerContext().Database(ctx, "db").Authenticator(ctx)
+
+			// Create a test user
+			user, err := a.NewUser(tc.UserName, "letmein", channels.BaseSetOf(t, "foo"))
+			require.NoError(t, err)
+			require.NoError(t, a.Save(user))
+
+			response := rt.Send(RequestByUser("PUT", "/db/AC+DC", `{"foo":"bar", "channels": ["foo"]}`, tc.UserName))
+			RequireStatus(t, response, 201)
+
+			response = rt.Send(RequestByUser("GET", "/db/_all_docs", "", tc.UserName))
+			RequireStatus(t, response, 200)
+
+			response = rt.Send(RequestByUser("GET", "/db/AC+DC", "", tc.UserName))
+			RequireStatus(t, response, 200)
+			var overlay struct {
+				Rev string `json:"_rev"`
+			}
+			require.NoError(t, json.Unmarshal(response.Body.Bytes(), &overlay))
+
+			response = rt.Send(RequestByUser("DELETE", "/db/AC+DC?rev="+overlay.Rev, "", tc.UserName))
+			RequireStatus(t, response, 200)
+		})
+	}
+}
+func TestRemovingUserXattr(t *testing.T) {
+	if base.UnitTestUrlIsWalrus() {
+		t.Skip("This test only works against Couchbase Server")
+	}
+
+	if !base.TestUseXattrs() {
+		t.Skip("This test only works with XATTRS enabled")
+	}
+
+	if !base.IsEnterpriseEdition() {
+		t.Skipf("test is EE only - user xattrs")
+	}
+
+	defer db.SuspendSequenceBatching()()
+
+	base.SetUpTestLogging(t, base.LevelDebug, base.KeyAll)
+
+	testCases := []struct {
+		name          string
+		autoImport    bool
+		importTrigger func(t *testing.T, rt *RestTester, docKey string)
+	}{
+		{
+			name:       "GET",
+			autoImport: false,
+			importTrigger: func(t *testing.T, rt *RestTester, docKey string) {
+				resp := rt.SendAdminRequest("GET", "/db/"+docKey, "")
+				RequireStatus(t, resp, http.StatusOK)
+			},
+		},
+		{
+			name:       "PUT",
+			autoImport: false,
+			importTrigger: func(t *testing.T, rt *RestTester, docKey string) {
+				resp := rt.SendAdminRequest("PUT", "/db/"+docKey, "{}")
+				RequireStatus(t, resp, http.StatusConflict)
+			},
+		},
+		{
+			name:       "Auto",
+			autoImport: true,
+			importTrigger: func(t *testing.T, rt *RestTester, docKey string) {
+				// No op
+			},
+		},
+	}
+
+	for _, testCase := range testCases {
+		t.Run(testCase.name, func(t *testing.T) {
+			docKey := testCase.name
+			xattrKey := "myXattr"
+			channelName := "testChan"
+
+			// Sync function to set channel access to whatever xattr is
+			rt := NewRestTester(t, &RestTesterConfig{
+				DatabaseConfig: &DatabaseConfig{DbConfig: DbConfig{
+					AutoImport:   testCase.autoImport,
+					UserXattrKey: xattrKey,
+				}},
+				SyncFn: `
+			function (doc, oldDoc, meta){
+				if (meta.xattrs.myXattr !== undefined){
+					channel(meta.xattrs.myXattr);
+				}
+			}`,
+			})
+
+			defer rt.Close()
+
+			gocbBucket, ok := base.AsUserXattrStore(rt.Bucket())
+			if !ok {
+				t.Skip("Test requires Couchbase Bucket")
+			}
+
+			// Initial PUT
+			resp := rt.SendAdminRequest("PUT", "/db/"+docKey, `{}`)
+			RequireStatus(t, resp, http.StatusCreated)
+
+			// Add xattr
+			_, err := gocbBucket.WriteUserXattr(docKey, xattrKey, channelName)
+			assert.NoError(t, err)
+
+			// Trigger import
+			testCase.importTrigger(t, rt, docKey)
+			err = rt.WaitForCondition(func() bool {
+				return rt.GetDatabase().DbStats.SharedBucketImport().ImportCount.Value() == 1
+			})
+			assert.NoError(t, err)
+
+			// Get sync data for doc and ensure user xattr has been used correctly to set channel
+			var syncData db.SyncData
+			subdocStore, ok := base.AsSubdocXattrStore(rt.Bucket())
+			require.True(t, ok)
+			_, err = subdocStore.SubdocGetXattr(docKey, base.SyncXattrName, &syncData)
+			assert.NoError(t, err)
+
+			assert.Equal(t, []string{channelName}, syncData.Channels.KeySet())
+
+			// Delete user xattr
+			_, err = gocbBucket.DeleteUserXattr(docKey, xattrKey)
+			assert.NoError(t, err)
+
+			// Trigger import
+			testCase.importTrigger(t, rt, docKey)
+			err = rt.WaitForCondition(func() bool {
+				return rt.GetDatabase().DbStats.SharedBucketImport().ImportCount.Value() == 2
+			})
+			assert.NoError(t, err)
+
+			// Ensure old channel set with user xattr has been removed
+			var syncData2 db.SyncData
+			_, err = subdocStore.SubdocGetXattr(docKey, base.SyncXattrName, &syncData2)
+			assert.NoError(t, err)
+
+			assert.Equal(t, uint64(3), syncData2.Channels[channelName].Seq)
+		})
+	}
+}
+func TestUserXattrAvoidRevisionIDGeneration(t *testing.T) {
+	if base.UnitTestUrlIsWalrus() {
+		t.Skip("This test only works against Couchbase Server")
+	}
+
+	if !base.TestUseXattrs() {
+		t.Skip("This test only works with XATTRS enabled")
+	}
+
+	if !base.IsEnterpriseEdition() {
+		t.Skipf("test is EE only - user xattrs")
+	}
+
+	base.SetUpTestLogging(t, base.LevelDebug, base.KeyAll)
+
+	docKey := t.Name()
+	xattrKey := "myXattr"
+	channelName := "testChan"
+
+	// Sync function to set channel access to whatever xattr is
+	rt := NewRestTester(t, &RestTesterConfig{
+		DatabaseConfig: &DatabaseConfig{DbConfig: DbConfig{
+			AutoImport:   true,
+			UserXattrKey: xattrKey,
+		}},
+		SyncFn: `
+			function (doc, oldDoc, meta){
+				if (meta.xattrs.myXattr !== undefined){
+					channel(meta.xattrs.myXattr);
+					console.log(JSON.stringify(meta));
+				}
+			}`,
+	})
+
+	defer rt.Close()
+
+	userXattrStore, ok := base.AsUserXattrStore(rt.Bucket())
+	if !ok {
+		t.Skip("Test requires Couchbase Bucket")
+	}
+
+	subdocXattrStore, ok := base.AsSubdocXattrStore(rt.Bucket())
+	require.True(t, ok)
+
+	// Initial PUT
+	resp := rt.SendAdminRequest("PUT", "/db/"+docKey, `{}`)
+	RequireStatus(t, resp, http.StatusCreated)
+
+	require.NoError(t, rt.WaitForPendingChanges())
+
+	// Get current sync data
+	var syncData db.SyncData
+	_, err := subdocXattrStore.SubdocGetXattr(docKey, base.SyncXattrName, &syncData)
+	assert.NoError(t, err)
+
+	docRev, err := rt.GetDatabase().GetSingleDatabaseCollection().GetRevisionCacheForTest().Get(base.TestCtx(t), docKey, syncData.CurrentRev, true, false)
+	assert.NoError(t, err)
+	assert.Equal(t, 0, len(docRev.Channels.ToArray()))
+	assert.Equal(t, syncData.CurrentRev, docRev.RevID)
+
+	// Write xattr to trigger import of user xattr
+	_, err = userXattrStore.WriteUserXattr(docKey, xattrKey, channelName)
+	assert.NoError(t, err)
+
+	// Wait for import
+	err = rt.WaitForCondition(func() bool {
+		return rt.GetDatabase().DbStats.SharedBucketImport().ImportCount.Value() == 1
+	})
+	assert.NoError(t, err)
+
+	// Ensure import worked and sequence incremented but that sequence did not
+	var syncData2 db.SyncData
+	_, err = subdocXattrStore.SubdocGetXattr(docKey, base.SyncXattrName, &syncData2)
+	assert.NoError(t, err)
+
+	docRev2, err := rt.GetDatabase().GetSingleDatabaseCollection().GetRevisionCacheForTest().Get(base.TestCtx(t), docKey, syncData.CurrentRev, true, false)
+	assert.NoError(t, err)
+	assert.Equal(t, syncData2.CurrentRev, docRev2.RevID)
+
+	assert.Equal(t, syncData.CurrentRev, syncData2.CurrentRev)
+	assert.True(t, syncData2.Sequence > syncData.Sequence)
+	assert.Equal(t, []string{channelName}, syncData2.Channels.KeySet())
+	assert.Equal(t, syncData2.Channels.KeySet(), docRev2.Channels.ToArray())
+
+	err = rt.Bucket().Set(docKey, 0, nil, []byte(`{"update": "update"}`))
+	assert.NoError(t, err)
+
+	err = rt.WaitForCondition(func() bool {
+		log.Printf("Import count is: %v", rt.GetDatabase().DbStats.SharedBucketImport().ImportCount.Value())
+		return rt.GetDatabase().DbStats.SharedBucketImport().ImportCount.Value() == 2
+	})
+	assert.NoError(t, err)
+
+	var syncData3 db.SyncData
+	_, err = subdocXattrStore.SubdocGetXattr(docKey, base.SyncXattrName, &syncData2)
+	assert.NoError(t, err)
+
+	assert.NotEqual(t, syncData2.CurrentRev, syncData3.CurrentRev)
 }
